refactor(test): extract more domains from ws-server.test.js (phase 2)
Extract 3 domain-focused test files from ws-server.test.js:
- ws-server-file-ops.test.js: directory listing, slash commands,
  agent listing, file browser symlink security, get_diff handler,
  browse_files and read_file handlers (~1250 lines)
- ws-server-permissions.test.js: POST /permission, auto permission
  mode confirmation, permission/question routing, POST /permission-response,
  request_session_context errors, reconnect permission recovery (~1300 lines)
- ws-server-broadcast.test.js: broadcastError, broadcastStatus,
  user_question_response forwarding, background session sync,
  broadcastToSession, public broadcast, agent idle/busy, models_updated,
  client_focus_changed, broadcast backpressure (~1200 lines)

Reduces ws-server.test.js from ~8114 to ~4153 lines.
All 270 tests pass across all ws-server*.test.js files.
