+++ packages/server/tests/ws-server.test.js
@@ -236,1066 +236,6 @@ describe('WsServer GET /health response shape', () => {
   })
 })
 
-describe('WsServer POST /permission with authRequired: false', () => {
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('accepts POST /permission without Bearer token when authRequired: false', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    // Connect a WebSocket client to receive the permission_request broadcast
-    const { ws, messages } = await createClient(port, true)
-
-    // Make a POST request to /permission WITHOUT Authorization header (don't await yet)
-    const responsePromise = fetch(`http://127.0.0.1:${port}/permission`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        tool_name: 'Bash',
-        tool_input: { command: 'ls -la' },
-      }),
-    })
-
-    // Wait for the permission_request broadcast
-    const permReq = await waitForMessage(messages, 'permission_request', 2000)
-    assert.ok(permReq, 'Should broadcast permission_request')
-    assert.equal(permReq.tool, 'Bash')
-
-    // Send permission response
-    send(ws, {
-      type: 'permission_response',
-      requestId: permReq.requestId,
-      decision: 'allow',
-    })
-
-    // Now await the HTTP response - it should complete
-    const response = await responsePromise
-
-    // Should return 200
-    assert.equal(response.status, 200, 'Should accept request without auth when authRequired: false')
-    assert.match(response.headers.get('content-type') ?? '', /^application\/json\b/, 'Response should be JSON')
-
-    const data = await response.json()
-    assert.equal(data.decision, 'allow', 'Should return the permission decision')
-
-    ws.close()
-  })
-
-  it('still rejects POST /permission without Bearer token when authRequired: true', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    // Make a POST request to /permission WITHOUT Authorization header
-    const response = await fetch(`http://127.0.0.1:${port}/permission`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        tool_name: 'Bash',
-        tool_input: { command: 'ls -la' },
-      }),
-    })
-
-    // Should return 403
-    assert.equal(response.status, 403, 'Should reject request without auth when authRequired: true')
-
-    const data = await response.json()
-    assert.equal(data.error, 'unauthorized')
-  })
-
-  it('accepts POST /permission with Bearer token when authRequired: true', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    // Connect client with auth using createClient helper (includes timeout/error handling)
-    const { ws: client, messages } = await createClient(port, false)
-
-    // Send auth
-    send(client, { type: 'auth', token: 'test-token' })
-
-    // Wait for auth_ok
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    // Make POST request WITH Authorization header (don't await yet)
-    const responsePromise = fetch(`http://127.0.0.1:${port}/permission`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'Authorization': 'Bearer test-token',
-      },
-      body: JSON.stringify({
-        tool_name: 'Bash',
-        tool_input: { command: 'ls -la' },
-      }),
-    })
-
-    // Verify we got the permission_request
-    const permReq = await waitForMessage(messages, 'permission_request', 2000)
-    assert.ok(permReq)
-
-    // Respond
-    send(client, {
-      type: 'permission_response',
-      requestId: permReq.requestId,
-      decision: 'deny',
-    })
-
-    // Now await the HTTP response
-    const response = await responsePromise
-
-    // Should return 200
-    assert.equal(response.status, 200, 'Should return 200 for request with valid Bearer token')
-
-    const data = await response.json()
-    assert.equal(data.decision, 'deny')
-
-    client.close()
-  })
-})
-
-describe('WsServer.broadcastError', () => {
-  let server
-  let port
-
-  beforeEach(() => {
-    port = 30000 + Math.floor(Math.random() * 10000)
-  })
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('broadcasts server_error to authenticated clients', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    server.start('127.0.0.1')
-    await once(server.httpServer, 'listening')
-
-    // Connect and auto-authenticate
-    const { ws, messages } = await createClient(port, true)
-
-    // Clear initial messages (auth_ok, server_mode, status, etc.)
-    messages.length = 0
-
-    // Broadcast a recoverable error
-    server.broadcastError('tunnel', 'Tunnel connection lost', true)
-
-    // Wait for the message
-    const errorMsg = await waitForMessage(messages, 'server_error', 1000)
-    assert.ok(errorMsg, 'Should receive server_error message')
-    assert.equal(errorMsg.type, 'server_error')
-    assert.equal(errorMsg.category, 'tunnel')
-    assert.equal(errorMsg.message, 'Tunnel connection lost')
-    assert.equal(errorMsg.recoverable, true)
-
-    ws.close()
-  })
-
-  it('broadcasts non-recoverable server_error', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    server.start('127.0.0.1')
-    await new Promise(r => setTimeout(r, 100))
-
-    const { ws, messages } = await createClient(port, true)
-    messages.length = 0
-
-    server.broadcastError('session', 'Process crashed', false)
-
-    const errorMsg = await waitForMessage(messages, 'server_error', 1000)
-    assert.equal(errorMsg.category, 'session')
-    assert.equal(errorMsg.message, 'Process crashed')
-    assert.equal(errorMsg.recoverable, false)
-
-    ws.close()
-  })
-
-  it('does not send server_error to unauthenticated clients', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: true,
-    })
-    server.start('127.0.0.1')
-    await once(server.httpServer, 'listening')
-
-    // Connect WITHOUT authenticating
-    const { ws, messages } = await createClient(port, false)
-
-    // Broadcast an error
-    server.broadcastError('general', 'Test error', true)
-
-    // Wait a bit to see if any message arrives
-    await new Promise(r => setTimeout(r, 200))
-
-    const errorMsg = messages.find(m => m.type === 'server_error')
-    assert.ok(!errorMsg, 'Unauthenticated client should not receive server_error')
-
-    ws.close()
-  })
-})
-
-describe('WsServer.broadcastStatus', () => {
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('broadcasts server_status to authenticated clients', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    // Connect and auto-authenticate
-    const { ws, messages } = await createClient(port, true)
-
-    // Clear initial messages (auth_ok, server_mode, status, etc.)
-    messages.length = 0
-
-    // Broadcast a status message
-    server.broadcastStatus('Tunnel recovered successfully')
-
-    // Wait for the message
-    const statusMsg = await waitForMessage(messages, 'server_status', 1000)
-    assert.ok(statusMsg, 'Should receive server_status message')
-    assert.equal(statusMsg.type, 'server_status')
-    assert.equal(statusMsg.message, 'Tunnel recovered successfully')
-
-    ws.close()
-  })
-
-  it('broadcasts multiple server_status messages', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, true)
-    messages.length = 0
-
-    // Broadcast multiple status messages
-    server.broadcastStatus('First recovery attempt')
-    server.broadcastStatus('Second recovery attempt')
-
-    // Wait until both server_status messages have been received
-    await waitFor(
-      () => messages.filter(m => m.type === 'server_status').length >= 2,
-      { timeoutMs: 2000, label: '2 server_status messages' }
-    )
-
-    const statusMsgs = messages.filter(m => m.type === 'server_status')
-    assert.equal(statusMsgs.length, 2, 'Should receive both status messages')
-    assert.equal(statusMsgs[0].message, 'First recovery attempt')
-    assert.equal(statusMsgs[1].message, 'Second recovery attempt')
-
-    ws.close()
-  })
-
-  it('does not send server_status to unauthenticated clients', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    // Connect WITHOUT authenticating
-    const { ws, messages } = await createClient(port, false)
-
-    // Broadcast a status message
-    server.broadcastStatus('Recovery in progress')
-
-    // Wait a bit to see if any message arrives
-    await new Promise(r => setTimeout(r, 200))
-
-    const statusMsg = messages.find(m => m.type === 'server_status')
-    assert.ok(!statusMsg, 'Unauthenticated client should not receive server_status')
-
-    ws.close()
-  })
-
-  it('broadcasts server_status to multiple authenticated clients', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    // Connect two authenticated clients
-    const client1 = await createClient(port, true)
-    const client2 = await createClient(port, true)
-
-    // Clear initial messages
-    client1.messages.length = 0
-    client2.messages.length = 0
-
-    // Broadcast status message
-    server.broadcastStatus('Both clients should receive this')
-
-    // Both clients should receive the message
-    const msg1 = await waitForMessage(client1.messages, 'server_status', 1000)
-    const msg2 = await waitForMessage(client2.messages, 'server_status', 1000)
-
-    assert.ok(msg1, 'Client 1 should receive server_status')
-    assert.ok(msg2, 'Client 2 should receive server_status')
-    assert.equal(msg1.message, 'Both clients should receive this')
-    assert.equal(msg2.message, 'Both clients should receive this')
-
-    client1.ws.close()
-    client2.ws.close()
-  })
-
-  it('includes correct message format in server_status', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, true)
-    messages.length = 0
-
-    // Broadcast with specific message
-    const testMessage = 'Connection re-established after timeout'
-    server.broadcastStatus(testMessage)
-
-    const statusMsg = await waitForMessage(messages, 'server_status', 1000)
-    assert.ok(statusMsg, 'Should receive server_status message')
-    assert.equal(typeof statusMsg.type, 'string', 'type should be a string')
-    assert.equal(statusMsg.type, 'server_status', 'type should be "server_status"')
-    assert.equal(typeof statusMsg.message, 'string', 'message should be a string')
-    assert.equal(statusMsg.message, testMessage, 'message should match the broadcast text')
-    assert.ok(!statusMsg.category, 'server_status should not have category field')
-    assert.ok(!statusMsg.recoverable, 'server_status should not have recoverable field')
-
-    ws.close()
-  })
-})
-
-describe('user_question_response forwarding (multi-session)', () => {
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('forwards user_question_response to active cli session', async () => {
-    const { manager, sessionsMap } = createMockSessionManager([
-      { id: 'sess-1', name: 'Test', cwd: '/tmp', type: 'cli' },
-    ])
-
-    const entry = sessionsMap.get('sess-1')
-    let receivedAnswer = null
-    entry.session.respondToQuestion = (answer) => { receivedAnswer = answer }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: manager,
-      defaultSessionId: 'sess-1',
-      authRequired: true,
-    })
-
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, false)
-
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    send(ws, { type: 'user_question_response', answer: 'Option A' })
-    await waitFor(() => receivedAnswer !== null, { label: 'receivedAnswer set' })
-
-    assert.equal(receivedAnswer, 'Option A', 'Answer should be forwarded to session')
-
-    ws.close()
-  })
-
-})
-
-describe('user_question_response forwarding (single-session)', () => {
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('forwards user_question_response to cliSession', async () => {
-    const mockSession = createMockSession()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    send(ws, { type: 'user_question_response', answer: 'Option A' })
-    await waitFor(() => mockSession.respondToQuestion.callCount >= 1, { label: 'respondToQuestion called' })
-
-    // Spy records calls — no manual tracking needed
-    assert.equal(mockSession.respondToQuestion.callCount, 1, 'respondToQuestion should be called once')
-    assert.deepStrictEqual(mockSession.respondToQuestion.lastCall, ['Option A', undefined], 'Answer should be forwarded to cliSession')
-
-    ws.close()
-  })
-
-  it('forwards answers map alongside text answer', async () => {
-    const mockSession = createMockSession()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    const answersMap = { 'Allow edit?': 'yes', 'Confirm?': 'no' }
-    send(ws, { type: 'user_question_response', answer: 'yes', answers: answersMap })
-    await waitFor(() => mockSession.respondToQuestion.callCount >= 1, { label: 'respondToQuestion called' })
-
-    assert.equal(mockSession.respondToQuestion.callCount, 1, 'respondToQuestion should be called once')
-    assert.deepStrictEqual(mockSession.respondToQuestion.lastCall, ['yes', answersMap], 'Both answer and answers map should be forwarded')
-
-    ws.close()
-  })
-
-  it('ignores user_question_response with non-string answer', async () => {
-    const mockSession = createMockSession()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    send(ws, { type: 'user_question_response', answer: 123 })
-    await new Promise(r => setTimeout(r, 100))
-
-    assert.equal(mockSession.respondToQuestion.callCount, 0, 'respondToQuestion should NOT be called for non-string answer')
-
-    ws.close()
-  })
-})
-
-describe('background session sync (_broadcastToSession)', () => {
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  /** Create a mock SessionManager with two sessions */
-  function createTwoSessionManager() {
-    const manager = new EventEmitter()
-    const sessionsMap = new Map()
-
-    const session1 = createMockSession()
-    session1.cwd = '/tmp/project-1'
-    sessionsMap.set('sess-1', { session: session1, name: 'Session 1', cwd: '/tmp/project-1', type: 'cli', isBusy: false })
-
-    const session2 = createMockSession()
-    session2.cwd = '/tmp/project-2'
-    sessionsMap.set('sess-2', { session: session2, name: 'Session 2', cwd: '/tmp/project-2', type: 'cli', isBusy: false })
-
-    manager.getSession = (id) => sessionsMap.get(id)
-    manager.listSessions = () => {
-      const list = []
-      for (const [id, entry] of sessionsMap) {
-        list.push({ sessionId: id, name: entry.name, cwd: entry.cwd, type: entry.type, isBusy: entry.isBusy })
-      }
-      return list
-    }
-    manager.getHistory = () => []
-    manager.recordUserInput = () => {}
-    manager.touchActivity = () => {}
-    manager.getFullHistoryAsync = async () => []
-    manager.isBudgetPaused = () => false
-    Object.defineProperty(manager, 'firstSessionId', {
-      get: () => sessionsMap.keys().next().value
-    })
-
-    return manager
-  }
-
-  it('tags session messages with sessionId', async () => {
-    const mockManager = createTwoSessionManager()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      defaultSessionId: 'sess-1',
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    // Emit a message event on session-1
-    mockManager.emit('session_event', {
-      sessionId: 'sess-1',
-      event: 'message',
-      data: { type: 'response', content: 'Hello from session 1', timestamp: Date.now() },
-    })
-
-    const msgEvent = await waitForMessageMatch(messages,
-      m => m.type === 'message' && m.content === 'Hello from session 1',
-      2000, 'message from session 1')
-    assert.ok(msgEvent, 'Should receive the message event')
-    assert.equal(msgEvent.sessionId, 'sess-1', 'Message should include sessionId')
-
-    ws.close()
-  })
-
-  it('does NOT deliver messages for sessions client is not viewing', async () => {
-    const mockManager = createTwoSessionManager()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      defaultSessionId: 'sess-1',
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    // Client is on sess-1, but we emit an event for sess-2
-    mockManager.emit('session_event', {
-      sessionId: 'sess-2',
-      event: 'message',
-      data: { type: 'response', content: 'Background message', timestamp: Date.now() },
-    })
-    await new Promise(r => setTimeout(r, 200))
-
-    const bgMsg = messages.find(m => m.type === 'message' && m.content === 'Background message')
-    assert.ok(!bgMsg, 'Client should NOT receive message for session it is not viewing')
-
-    ws.close()
-  })
-
-  it('does NOT deliver stream_start/stream_end for non-viewed sessions', async () => {
-    const mockManager = createTwoSessionManager()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      defaultSessionId: 'sess-1',
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    // Emit stream lifecycle for sess-2 while client is on sess-1
-    mockManager.emit('session_event', {
-      sessionId: 'sess-2',
-      event: 'stream_start',
-      data: { messageId: 'msg-bg-1' },
-    })
-    mockManager.emit('session_event', {
-      sessionId: 'sess-2',
-      event: 'stream_end',
-      data: { messageId: 'msg-bg-1' },
-    })
-    await new Promise(r => setTimeout(r, 200))
-
-    const streamStart = messages.find(m => m.type === 'stream_start' && m.messageId === 'msg-bg-1')
-    assert.ok(!streamStart, 'Should NOT receive stream_start for non-viewed session')
-
-    const streamEnd = messages.find(m => m.type === 'stream_end' && m.messageId === 'msg-bg-1')
-    assert.ok(!streamEnd, 'Should NOT receive stream_end for non-viewed session')
-
-    ws.close()
-  })
-
-  it('scopes model_changed and permission_mode_changed to active session (#1138)', async () => {
-    const mockManager = createTwoSessionManager()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      defaultSessionId: 'sess-1',
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    // Emit a "ready" session_event for sess-2 (client is on sess-1). The
-    // EventNormalizer translates ready into model_changed + permission_mode_changed
-    // messages, which are broadcast via broadcastToSession with the default
-    // activeSessionId filter — so the client on sess-1 should not receive them.
-    mockManager.emit('session_event', {
-      sessionId: 'sess-2',
-      event: 'ready',
-      data: {},
-    })
-    await new Promise(r => setTimeout(r, 200))
-
-    const modelMsg = messages.find(m => m.type === 'model_changed' && m.sessionId === 'sess-2')
-    assert.ok(!modelMsg, 'Should NOT receive model_changed for non-viewed session')
-
-    const permMsg = messages.find(m => m.type === 'permission_mode_changed' && m.sessionId === 'sess-2')
-    assert.ok(!permMsg, 'Should NOT receive permission_mode_changed for non-viewed session')
-
-    ws.close()
-  })
-
-})
-
-describe('broadcastToSession via message-handler path (#1344)', () => {
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  function createTwoSessionManager() {
-    const manager = new EventEmitter()
-    const sessionsMap = new Map()
-
-    const session1 = createMockSession()
-    session1.cwd = '/tmp/project-1'
-    sessionsMap.set('sess-1', { session: session1, name: 'Session 1', cwd: '/tmp/project-1', type: 'cli', isBusy: false })
-
-    const session2 = createMockSession()
-    session2.cwd = '/tmp/project-2'
-    sessionsMap.set('sess-2', { session: session2, name: 'Session 2', cwd: '/tmp/project-2', type: 'cli', isBusy: false })
-
-    manager.getSession = (id) => sessionsMap.get(id)
-    manager.listSessions = () => {
-      const list = []
-      for (const [id, entry] of sessionsMap) {
-        list.push({ sessionId: id, name: entry.name, cwd: entry.cwd, type: entry.type, isBusy: entry.isBusy })
-      }
-      return list
-    }
-    manager.getHistory = () => []
-    manager.recordUserInput = () => {}
-    manager.touchActivity = () => {}
-    manager.getFullHistoryAsync = async () => []
-    manager.isBudgetPaused = () => false
-    manager.getSessionContext = async () => null
-    Object.defineProperty(manager, 'firstSessionId', {
-      get: () => sessionsMap.keys().next().value
-    })
-
-    return manager
-  }
-
-  it('set_model broadcasts model_changed only to clients on the target session', async () => {
-    const manager = createTwoSessionManager()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: manager,
-      defaultSessionId: 'sess-1',
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    // Client A: stays on sess-1 (default)
-    const clientA = await createClient(port, false)
-    send(clientA.ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(clientA.messages, 'auth_ok', 2000)
-
-    // Client B: switches to sess-2
-    const clientB = await createClient(port, false)
-    send(clientB.ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(clientB.messages, 'auth_ok', 2000)
-    send(clientB.ws, { type: 'switch_session', sessionId: 'sess-2' })
-    await waitForMessage(clientB.messages, 'session_switched', 2000)
-
-    // Client B sends set_model targeting sess-2
-    send(clientB.ws, { type: 'set_model', model: 'sonnet', sessionId: 'sess-2' })
-
-    // Client B (on sess-2) should receive model_changed tagged with sess-2
-    // (filter by sessionId to distinguish from _sendSessionInfo's untagged model_changed)
-    const bModelMsg = await waitForMessageMatch(clientB.messages,
-      m => m.type === 'model_changed' && m.sessionId === 'sess-2',
-      2000, 'model_changed for sess-2')
-    assert.ok(bModelMsg, 'Client on sess-2 should receive model_changed with sessionId tag')
-    assert.equal(bModelMsg.model, 'sonnet', 'model_changed should contain the new model')
-
-    // Client A (on sess-1) should NOT receive model_changed for sess-2
-    const aModelMsg = clientA.messages.find(m => m.type === 'model_changed' && m.sessionId === 'sess-2')
-    assert.ok(!aModelMsg, 'Client on sess-1 should NOT receive model_changed for sess-2')
-
-    clientA.ws.close()
-    clientB.ws.close()
-  })
-
-  it('set_permission_mode broadcasts permission_mode_changed only to clients on the target session', async () => {
-    const manager = createTwoSessionManager()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: manager,
-      defaultSessionId: 'sess-1',
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    // Client A: stays on sess-1 (default)
-    const clientA = await createClient(port, false)
-    send(clientA.ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(clientA.messages, 'auth_ok', 2000)
-
-    // Client B: switches to sess-2
-    const clientB = await createClient(port, false)
-    send(clientB.ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(clientB.messages, 'auth_ok', 2000)
-    send(clientB.ws, { type: 'switch_session', sessionId: 'sess-2' })
-    await waitForMessage(clientB.messages, 'session_switched', 2000)
-
-    // Client B sends set_permission_mode targeting sess-2
-    send(clientB.ws, { type: 'set_permission_mode', mode: 'approve', sessionId: 'sess-2' })
-
-    // Client B (on sess-2) should receive permission_mode_changed tagged with sess-2
-    // (filter by sessionId to distinguish from _sendSessionInfo's untagged permission_mode_changed)
-    const bPermMsg = await waitForMessageMatch(clientB.messages,
-      m => m.type === 'permission_mode_changed' && m.sessionId === 'sess-2',
-      2000, 'permission_mode_changed for sess-2')
-    assert.ok(bPermMsg, 'Client on sess-2 should receive permission_mode_changed with sessionId tag')
-    assert.equal(bPermMsg.mode, 'approve', 'permission_mode_changed should contain the new mode')
-
-    // Client A (on sess-1) should NOT receive permission_mode_changed for sess-2
-    const aPermMsg = clientA.messages.find(m => m.type === 'permission_mode_changed' && m.sessionId === 'sess-2')
-    assert.ok(!aPermMsg, 'Client on sess-1 should NOT receive permission_mode_changed for sess-2')
-
-    clientA.ws.close()
-    clientB.ws.close()
-  })
-})
-
-describe('public broadcast method', () => {
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('broadcast() sends to all authenticated clients', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    // Use public broadcast to send discovered_sessions
-    server.broadcast({ type: 'discovered_sessions', tmux: [{ sessionName: 'test-session', cwd: '/tmp', pid: 123 }] })
-
-    const discoveryMsg = await waitForMessage(messages, 'discovered_sessions')
-    assert.ok(discoveryMsg, 'Client should receive broadcast message')
-    assert.equal(discoveryMsg.tmux.length, 1)
-    assert.equal(discoveryMsg.tmux[0].sessionName, 'test-session')
-
-    ws.close()
-  })
-
-  it('broadcast() does not send to unauthenticated clients', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    // Connect but do NOT authenticate
-    const { ws, messages } = await createClient(port, false)
-    await new Promise(r => setTimeout(r, 100))
-
-    server.broadcast({ type: 'discovered_sessions', tmux: [{ sessionName: 'test-session', cwd: '/tmp', pid: 123 }] })
-    await new Promise(r => setTimeout(r, 100))
-
-    const discoveryMsg = messages.find(m => m.type === 'discovered_sessions')
-    assert.ok(!discoveryMsg, 'Unauthenticated client should NOT receive broadcast')
-
-    ws.close()
-  })
-})
-
-describe('agent idle/busy notifications', () => {
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  /** Create a mock SessionManager with two sessions */
-  function createTwoSessionManager() {
-    const manager = new EventEmitter()
-    const sessionsMap = new Map()
-
-    const session1 = createMockSession()
-    session1.cwd = '/tmp/project-1'
-    sessionsMap.set('sess-1', { session: session1, name: 'Session 1', cwd: '/tmp/project-1', type: 'cli', isBusy: false })
-
-    const session2 = createMockSession()
-    session2.cwd = '/tmp/project-2'
-    sessionsMap.set('sess-2', { session: session2, name: 'Session 2', cwd: '/tmp/project-2', type: 'cli', isBusy: false })
-
-    manager.getSession = (id) => sessionsMap.get(id)
-    manager.listSessions = () => {
-      const list = []
-      for (const [id, entry] of sessionsMap) {
-        list.push({ sessionId: id, name: entry.name, cwd: entry.cwd, type: entry.type, isBusy: entry.isBusy })
-      }
-      return list
-    }
-    manager.getHistory = () => []
-    manager.recordUserInput = () => {}
-    manager.touchActivity = () => {}
-    manager.getFullHistoryAsync = async () => []
-    manager.isBudgetPaused = () => false
-    manager.getSessionContext = async () => null
-    Object.defineProperty(manager, 'firstSessionId', {
-      get: () => sessionsMap.keys().next().value
-    })
-
-    return manager
-  }
-
-  it('broadcasts agent_busy after stream_start', async () => {
-    const mockManager = createTwoSessionManager()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      defaultSessionId: 'sess-1',
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    messages.length = 0
-
-    mockManager.emit('session_event', {
-      sessionId: 'sess-1',
-      event: 'stream_start',
-      data: { messageId: 'msg-1' },
-    })
-
-    const agentBusy = await waitForMessage(messages, 'agent_busy')
-    assert.ok(agentBusy, 'Should receive agent_busy after stream_start')
-    assert.equal(agentBusy.sessionId, 'sess-1', 'agent_busy should include sessionId')
-
-    ws.close()
-  })
-
-  it('broadcasts agent_idle after result', async () => {
-    const mockManager = createTwoSessionManager()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      defaultSessionId: 'sess-1',
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    messages.length = 0
-
-    mockManager.emit('session_event', {
-      sessionId: 'sess-1',
-      event: 'result',
-      data: { cost: 0.01, duration: 500, usage: {} },
-    })
-
-    const agentIdle = await waitForMessage(messages, 'agent_idle')
-    assert.ok(agentIdle, 'Should receive agent_idle after result')
-    assert.equal(agentIdle.sessionId, 'sess-1', 'agent_idle should include sessionId')
-
-    ws.close()
-  })
-
-  it('broadcasts session_list on stream_start for immediate busy dot', async () => {
-    const mockManager = createTwoSessionManager()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      defaultSessionId: 'sess-1',
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    messages.length = 0
-
-    mockManager.emit('session_event', {
-      sessionId: 'sess-1',
-      event: 'stream_start',
-      data: { messageId: 'msg-1' },
-    })
-
-    const sessionList = await waitForMessage(messages, 'session_list')
-    assert.ok(sessionList, 'Should receive session_list broadcast on stream_start')
-    assert.ok(Array.isArray(sessionList.sessions), 'session_list should contain sessions array')
-
-    ws.close()
-  })
-
-  it('hasActiveViewersForSession returns correct values', async () => {
-    const mockManager = createTwoSessionManager()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      defaultSessionId: 'sess-1',
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    // No clients — should be false for both
-    assert.equal(server.hasActiveViewersForSession('sess-1'), false, 'No viewers when no clients')
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    // Client is on sess-1 (default)
-    assert.equal(server.hasActiveViewersForSession('sess-1'), true, 'Client is viewing sess-1')
-    assert.equal(server.hasActiveViewersForSession('sess-2'), false, 'No client is viewing sess-2')
-
-    ws.close()
-  })
-})
 
 describe('WsServer drain behavior (multi-session mode)', () => {
   let server
@@ -1846,1689 +786,6 @@ describe('primary client tracking', () => {
   })
 })
 
-describe('auto permission mode confirmation handshake', () => {
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('auto mode requires confirmation (single-session)', async () => {
-    const mockSession = createMockSession()
-    let appliedMode = null
-    mockSession.setPermissionMode = (mode) => { appliedMode = mode }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-    messages.length = 0
-
-    // Request auto mode without confirmed flag
-    send(ws, { type: 'set_permission_mode', mode: 'auto' })
-
-    // Should get confirm_permission_mode challenge
-    const confirm = await waitForMessage(messages, 'confirm_permission_mode', 2000)
-    assert.ok(confirm, 'Should receive confirm_permission_mode')
-    assert.equal(confirm.mode, 'auto')
-    assert.equal(typeof confirm.warning, 'string')
-    assert.ok(confirm.warning.length > 0)
-
-    // Mode should NOT have been applied
-    assert.equal(appliedMode, null, 'Auto mode should NOT be applied without confirmation')
-
-    // No permission_mode_changed should have been broadcast
-    const modeChanged = messages.find(m => m.type === 'permission_mode_changed')
-    assert.equal(modeChanged, undefined, 'permission_mode_changed should NOT be sent')
-
-    ws.close()
-  })
-
-  it('confirmed auto mode applies normally (single-session)', async () => {
-    const mockSession = createMockSession()
-    let appliedMode = null
-    mockSession.setPermissionMode = (mode) => { appliedMode = mode }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-    messages.length = 0
-
-    // Send with confirmed: true
-    send(ws, { type: 'set_permission_mode', mode: 'auto', confirmed: true })
-
-    const modeChanged = await waitForMessage(messages, 'permission_mode_changed', 2000)
-    assert.ok(modeChanged, 'Should receive permission_mode_changed')
-    assert.equal(modeChanged.mode, 'auto')
-    assert.equal(appliedMode, 'auto', 'Auto mode should be applied with confirmation')
-
-    ws.close()
-  })
-
-  it('non-auto modes bypass confirmation (single-session)', async () => {
-    const mockSession = createMockSession()
-    let appliedMode = null
-    mockSession.setPermissionMode = (mode) => { appliedMode = mode }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-    messages.length = 0
-
-    // Send approve mode (no confirmation needed)
-    send(ws, { type: 'set_permission_mode', mode: 'approve' })
-
-    const modeChanged = await waitForMessage(messages, 'permission_mode_changed', 2000)
-    assert.ok(modeChanged, 'Should receive permission_mode_changed immediately')
-    assert.equal(modeChanged.mode, 'approve')
-    assert.equal(appliedMode, 'approve')
-
-    // No confirm_permission_mode should have been sent
-    const confirm = messages.find(m => m.type === 'confirm_permission_mode')
-    assert.equal(confirm, undefined, 'Non-auto modes should not trigger confirmation')
-
-    ws.close()
-  })
-
-  it('acceptEdits mode bypasses confirmation (single-session)', async () => {
-    const mockSession = createMockSession()
-    let appliedMode = null
-    mockSession.setPermissionMode = (mode) => { appliedMode = mode }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-    messages.length = 0
-
-    // Send acceptEdits mode (no confirmation needed — safer than auto)
-    send(ws, { type: 'set_permission_mode', mode: 'acceptEdits' })
-
-    const modeChanged = await waitForMessage(messages, 'permission_mode_changed', 2000)
-    assert.ok(modeChanged, 'Should receive permission_mode_changed immediately')
-    assert.equal(modeChanged.mode, 'acceptEdits')
-    assert.equal(appliedMode, 'acceptEdits')
-
-    // No confirm_permission_mode should have been sent
-    const confirm = messages.find(m => m.type === 'confirm_permission_mode')
-    assert.equal(confirm, undefined, 'acceptEdits should not trigger confirmation')
-
-    ws.close()
-  })
-
-  it('auto mode requires confirmation (multi-session)', async () => {
-    const manager = new EventEmitter()
-    const mockSession = createMockSession()
-    let appliedMode = null
-    mockSession.setPermissionMode = (mode) => { appliedMode = mode }
-    mockSession.cwd = '/tmp/test'
-
-    const sessionsMap = new Map()
-    sessionsMap.set('sess-1', { session: mockSession, name: 'Test', cwd: '/tmp/test', type: 'cli', isBusy: false })
-    manager.getSession = (id) => sessionsMap.get(id)
-    manager.listSessions = () => [{ id: 'sess-1', name: 'Test', cwd: '/tmp/test', type: 'cli', isBusy: false }]
-    manager.getHistory = () => []
-    manager.recordUserInput = () => {}
-    manager.getFullHistoryAsync = async () => []
-    manager.isBudgetPaused = () => false
-    Object.defineProperty(manager, 'firstSessionId', { get: () => 'sess-1' })
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: manager,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-    messages.length = 0
-
-    send(ws, { type: 'set_permission_mode', mode: 'auto' })
-
-    const confirm = await waitForMessage(messages, 'confirm_permission_mode', 2000)
-    assert.ok(confirm, 'Should receive confirm_permission_mode in multi-session mode')
-    assert.equal(confirm.mode, 'auto')
-    assert.equal(appliedMode, null, 'Mode should not be applied without confirmation')
-
-    ws.close()
-  })
-})
-
-describe('directory listing', () => {
-  let server
-  const TOKEN = 'test-token'
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('lists directories at a valid path', async () => {
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      cliSession: createMockSession(),
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: TOKEN })
-    await waitForMessage(messages, 'auth_ok', 2000)
-    messages.length = 0
-
-    // List home directory — should always exist and contain directories
-    send(ws, { type: 'list_directory', path: '~' })
-
-    const listing = await waitForMessage(messages, 'directory_listing', 2000)
-    assert.ok(listing, 'Should receive directory_listing')
-    assert.equal(listing.error, null)
-    assert.ok(Array.isArray(listing.entries))
-
-    ws.close()
-  })
-
-  it('returns error for non-existent path', async () => {
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      cliSession: createMockSession(),
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: TOKEN })
-    await waitForMessage(messages, 'auth_ok', 2000)
-    messages.length = 0
-
-    // Use a path inside the home directory that doesn't exist
-    const os = await import('os')
-    const nonexistent = `${os.homedir()}/nonexistent_path_that_does_not_exist_12345`
-    send(ws, { type: 'list_directory', path: nonexistent })
-
-    const listing = await waitForMessage(messages, 'directory_listing', 2000)
-    assert.ok(listing, 'Should receive directory_listing')
-    assert.equal(listing.error, 'Directory not found')
-    assert.deepEqual(listing.entries, [])
-
-    ws.close()
-  })
-
-  it('returns error for a file path', async () => {
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      cliSession: createMockSession(),
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: TOKEN })
-    await waitForMessage(messages, 'auth_ok', 2000)
-    messages.length = 0
-
-    // Use this test file itself as a file path
-    const filePath = new URL(import.meta.url).pathname
-    send(ws, { type: 'list_directory', path: filePath })
-
-    const listing = await waitForMessage(messages, 'directory_listing', 2000)
-    assert.ok(listing, 'Should receive directory_listing')
-    assert.equal(listing.error, 'Not a directory')
-    assert.deepEqual(listing.entries, [])
-
-    ws.close()
-  })
-
-  it('filters hidden directories', async () => {
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      cliSession: createMockSession(),
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: TOKEN })
-    await waitForMessage(messages, 'auth_ok', 2000)
-    messages.length = 0
-
-    // List home directory — should have entries but none starting with '.'
-    send(ws, { type: 'list_directory', path: '~' })
-
-    const listing = await waitForMessage(messages, 'directory_listing', 2000)
-    assert.ok(listing, 'Should receive directory_listing')
-    assert.equal(listing.error, null)
-    const hidden = listing.entries.filter(e => e.name.startsWith('.'))
-    assert.equal(hidden.length, 0, 'Should not include hidden directories')
-
-    ws.close()
-  })
-
-  it('requires authentication', async () => {
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      cliSession: createMockSession(),
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, false)
-
-    // Send list_directory before authenticating
-    send(ws, { type: 'list_directory', path: '~' })
-    await new Promise(r => setTimeout(r, 200))
-
-    // Should NOT get any directory_listing back (message is ignored pre-auth)
-    const listing = messages.find(m => m.type === 'directory_listing')
-    assert.equal(listing, undefined, 'Should not respond to unauthenticated requests')
-
-    ws.close()
-  })
-
-  it('defaults to home directory when path is empty', async () => {
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      cliSession: createMockSession(),
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: TOKEN })
-    await waitForMessage(messages, 'auth_ok', 2000)
-    messages.length = 0
-
-    send(ws, { type: 'list_directory' })
-
-    const listing = await waitForMessage(messages, 'directory_listing', 2000)
-    assert.ok(listing, 'Should receive directory_listing')
-    assert.equal(listing.error, null)
-    assert.ok(listing.path, 'Should have a resolved path')
-    assert.ok(listing.entries.length > 0, 'Home directory should have entries')
-
-    ws.close()
-  })
-
-  it('works in multi-session mode', async () => {
-    const manager = new EventEmitter()
-    const mockSession = createMockSession()
-    mockSession.cwd = '/tmp/test'
-
-    const sessionsMap = new Map()
-    sessionsMap.set('sess-1', { session: mockSession, name: 'Test', cwd: '/tmp/test', type: 'cli', isBusy: false })
-    manager.getSession = (id) => sessionsMap.get(id)
-    manager.listSessions = () => [{ id: 'sess-1', name: 'Test', cwd: '/tmp/test', type: 'cli', isBusy: false }]
-    manager.getHistory = () => []
-    manager.recordUserInput = () => {}
-    manager.getFullHistoryAsync = async () => []
-    manager.isBudgetPaused = () => false
-    Object.defineProperty(manager, 'firstSessionId', { get: () => 'sess-1' })
-
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      sessionManager: manager,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: TOKEN })
-    await waitForMessage(messages, 'auth_ok', 2000)
-    messages.length = 0
-
-    send(ws, { type: 'list_directory', path: '~' })
-
-    const listing = await waitForMessage(messages, 'directory_listing', 2000)
-    assert.ok(listing, 'Should receive directory_listing in multi-session mode')
-    assert.equal(listing.error, null)
-
-    ws.close()
-  })
-
-  it('restricts listing to home directory', async () => {
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      cliSession: createMockSession(),
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, false)
-    send(ws, { type: 'auth', token: TOKEN })
-    await waitForMessage(messages, 'auth_ok', 2000)
-    messages.length = 0
-
-    // Try listing /tmp — should be denied (outside home directory)
-    send(ws, { type: 'list_directory', path: '/tmp' })
-
-    const listing = await waitForMessage(messages, 'directory_listing', 2000)
-    assert.ok(listing, 'Should receive directory_listing')
-    assert.ok(listing.error.includes('restricted'), 'Should get access denied error')
-    assert.deepEqual(listing.entries, [])
-
-    ws.close()
-  })
-
-  it('rejects symlink inside home that points outside home (#662)', async () => {
-    // Create a temp directory inside home with a symlink escaping to /tmp
-    const home = homedir()
-    const testDir = mkdtempSync(join(home, '.chroxy-test-symlink-'))
-    const outsideTarget = mkdtempSync(join(tmpdir(), 'chroxy-test-outside-'))
-    writeFileSync(join(outsideTarget, 'leaked.txt'), 'should not see this')
-    mkdirSync(join(outsideTarget, 'leaked-dir'))
-
-    try {
-      symlinkSync(outsideTarget, join(testDir, 'escape-link'))
-
-      server = new WsServer({
-        port: 0,
-        apiToken: TOKEN,
-        cliSession: createMockSession(),
-        authRequired: true,
-      })
-      const port = await startServerAndGetPort(server)
-      const { ws, messages } = await createClient(port, false)
-      send(ws, { type: 'auth', token: TOKEN })
-      await waitForMessage(messages, 'auth_ok', 2000)
-      messages.length = 0
-
-      // Try listing through the symlink — should be denied
-      send(ws, { type: 'list_directory', path: join(testDir, 'escape-link') })
-
-      const listing = await waitForMessage(messages, 'directory_listing', 2000)
-      assert.ok(listing, 'Should receive directory_listing')
-      assert.ok(listing.error, 'Should return an error for symlink outside home')
-      assert.match(listing.error, /restricted/i)
-      assert.deepEqual(listing.entries, [])
-
-      ws.close()
-    } finally {
-      rmSync(testDir, { recursive: true, force: true })
-      rmSync(outsideTarget, { recursive: true, force: true })
-    }
-  })
-
-  it('allows symlink inside home that points within home (#662)', async () => {
-    // Create a temp directory inside home with a symlink pointing to another dir in home
-    const home = homedir()
-    const testDir = mkdtempSync(join(home, '.chroxy-test-symlink-'))
-    const internalTarget = join(testDir, 'real-dir')
-    mkdirSync(internalTarget)
-    mkdirSync(join(internalTarget, 'child'))
-
-    try {
-      symlinkSync(internalTarget, join(testDir, 'internal-link'))
-
-      server = new WsServer({
-        port: 0,
-        apiToken: TOKEN,
-        cliSession: createMockSession(),
-        authRequired: true,
-      })
-      const port = await startServerAndGetPort(server)
-      const { ws, messages } = await createClient(port, false)
-      send(ws, { type: 'auth', token: TOKEN })
-      await waitForMessage(messages, 'auth_ok', 2000)
-      messages.length = 0
-
-      // List through the symlink — should work since target is inside home
-      send(ws, { type: 'list_directory', path: join(testDir, 'internal-link') })
-
-      const listing = await waitForMessage(messages, 'directory_listing', 2000)
-      assert.ok(listing, 'Should receive directory_listing')
-      assert.equal(listing.error, null, 'Should not return error for symlink within home')
-      assert.ok(listing.entries.some(e => e.name === 'child'), 'Should list child directory')
-
-      ws.close()
-    } finally {
-      rmSync(testDir, { recursive: true, force: true })
-    }
-  })
-})
-
-describe('slash commands', () => {
-  let server
-  const TOKEN = 'test-token'
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('returns commands from project .claude/commands/ directory', async () => {
-    const { mkdirSync, writeFileSync, rmSync } = await import('fs')
-    const { join } = await import('path')
-    const { tmpdir } = await import('os')
-
-    // Create temp project with .claude/commands/
-    const tmpDir = join(tmpdir(), `chroxy-test-slash-${Date.now()}`)
-    const cmdDir = join(tmpDir, '.claude', 'commands')
-    mkdirSync(cmdDir, { recursive: true })
-    writeFileSync(join(cmdDir, 'deploy.md'), '# /deploy\n\nDeploy to production.\n\n## Steps\n...')
-    writeFileSync(join(cmdDir, 'test.md'), '# /test\n\nRun the test suite.\n')
-
-    try {
-      const mockSession = createMockSession()
-      mockSession.cwd = tmpDir
-
-      server = new WsServer({
-        port: 0,
-        apiToken: TOKEN,
-        cliSession: mockSession,
-        authRequired: true,
-      })
-      const port = await startServerAndGetPort(server)
-      const { ws, messages } = await createClient(port, false)
-      send(ws, { type: 'auth', token: TOKEN })
-      await waitForMessage(messages, 'auth_ok', 2000)
-      messages.length = 0
-
-      send(ws, { type: 'list_slash_commands' })
-      const result = await waitForMessage(messages, 'slash_commands', 2000)
-
-      assert.ok(result, 'Should receive slash_commands')
-      assert.ok(Array.isArray(result.commands))
-      assert.ok(result.commands.length >= 2, 'Should find at least 2 commands')
-
-      const deploy = result.commands.find(c => c.name === 'deploy')
-      assert.ok(deploy, 'Should include deploy command')
-      assert.equal(deploy.source, 'project')
-      assert.ok(deploy.description.length > 0, 'Should extract description')
-
-      const test = result.commands.find(c => c.name === 'test')
-      assert.ok(test, 'Should include test command')
-
-      ws.close()
-    } finally {
-      rmSync(tmpDir, { recursive: true, force: true })
-    }
-  })
-
-  it('returns empty array when no commands exist', async () => {
-    const { mkdirSync, rmSync } = await import('fs')
-    const { join } = await import('path')
-    const { tmpdir } = await import('os')
-
-    const tmpDir = join(tmpdir(), `chroxy-test-slash-empty-${Date.now()}`)
-    mkdirSync(tmpDir, { recursive: true })
-
-    try {
-      const mockSession = createMockSession()
-      mockSession.cwd = tmpDir
-
-      server = new WsServer({
-        port: 0,
-        apiToken: TOKEN,
-        cliSession: mockSession,
-        authRequired: true,
-      })
-      const port = await startServerAndGetPort(server)
-      const { ws, messages } = await createClient(port, false)
-      send(ws, { type: 'auth', token: TOKEN })
-      await waitForMessage(messages, 'auth_ok', 2000)
-      messages.length = 0
-
-      send(ws, { type: 'list_slash_commands' })
-      const result = await waitForMessage(messages, 'slash_commands', 2000)
-
-      assert.ok(result, 'Should receive slash_commands')
-      assert.ok(Array.isArray(result.commands))
-
-      ws.close()
-    } finally {
-      rmSync(tmpDir, { recursive: true, force: true })
-    }
-  })
-
-  it('works in multi-session mode', async () => {
-    const { mkdirSync, writeFileSync, rmSync } = await import('fs')
-    const { join } = await import('path')
-    const { tmpdir } = await import('os')
-
-    const tmpDir = join(tmpdir(), `chroxy-test-slash-ms-${Date.now()}`)
-    const cmdDir = join(tmpDir, '.claude', 'commands')
-    mkdirSync(cmdDir, { recursive: true })
-    writeFileSync(join(cmdDir, 'build.md'), '# /build\n\nBuild the project.')
-
-    try {
-      const manager = new EventEmitter()
-      const mockSession = createMockSession()
-      mockSession.cwd = tmpDir
-
-      const sessionsMap = new Map()
-      sessionsMap.set('sess-1', { session: mockSession, name: 'Test', cwd: tmpDir, type: 'cli', isBusy: false })
-      manager.getSession = (id) => sessionsMap.get(id)
-      manager.listSessions = () => [{ id: 'sess-1', name: 'Test', cwd: tmpDir, type: 'cli', isBusy: false }]
-      manager.getHistory = () => []
-      manager.recordUserInput = () => {}
-      manager.getFullHistoryAsync = async () => []
-      manager.isBudgetPaused = () => false
-      Object.defineProperty(manager, 'firstSessionId', { get: () => 'sess-1' })
-
-      server = new WsServer({
-        port: 0,
-        apiToken: TOKEN,
-        sessionManager: manager,
-        authRequired: true,
-      })
-      const port = await startServerAndGetPort(server)
-      const { ws, messages } = await createClient(port, false)
-      send(ws, { type: 'auth', token: TOKEN })
-      await waitForMessage(messages, 'auth_ok', 2000)
-      messages.length = 0
-
-      send(ws, { type: 'list_slash_commands' })
-      const result = await waitForMessage(messages, 'slash_commands', 2000)
-
-      assert.ok(result, 'Should receive slash_commands in multi-session mode')
-      assert.equal(result.sessionId, 'sess-1', 'slash_commands should include sessionId in multi-session mode')
-      const build = result.commands.find(c => c.name === 'build')
-      assert.ok(build, 'Should include build command')
-      assert.equal(build.source, 'project')
-
-      ws.close()
-    } finally {
-      rmSync(tmpDir, { recursive: true, force: true })
-    }
-  })
-
-  it('omits sessionId in single-session CLI mode', async () => {
-    const { mkdirSync, rmSync } = await import('fs')
-    const { join } = await import('path')
-    const { tmpdir } = await import('os')
-
-    const tmpDir = join(tmpdir(), `chroxy-test-slash-cli-${Date.now()}`)
-    mkdirSync(tmpDir, { recursive: true })
-
-    try {
-      const mockSession = createMockSession()
-      mockSession.cwd = tmpDir
-
-      server = new WsServer({
-        port: 0,
-        apiToken: TOKEN,
-        cliSession: mockSession,
-        authRequired: true,
-      })
-      const port = await startServerAndGetPort(server)
-      const { ws, messages } = await createClient(port, false)
-      send(ws, { type: 'auth', token: TOKEN })
-      await waitForMessage(messages, 'auth_ok', 2000)
-      messages.length = 0
-
-      send(ws, { type: 'list_slash_commands' })
-      const result = await waitForMessage(messages, 'slash_commands', 2000)
-
-      assert.ok(result, 'Should receive slash_commands')
-      assert.equal(result.sessionId, undefined, 'slash_commands should NOT include sessionId in single-session mode')
-
-      ws.close()
-    } finally {
-      rmSync(tmpDir, { recursive: true, force: true })
-    }
-  })
-})
-
-describe('agent listing', () => {
-  let server
-  const TOKEN = 'test-token'
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('returns agents from project .claude/agents/ directory', async () => {
-    const { mkdirSync, writeFileSync, rmSync } = await import('fs')
-    const { join } = await import('path')
-    const { tmpdir } = await import('os')
-
-    const tmpDir = join(tmpdir(), `chroxy-test-agents-${Date.now()}`)
-    const agentDir = join(tmpDir, '.claude', 'agents')
-    mkdirSync(agentDir, { recursive: true })
-    writeFileSync(join(agentDir, 'reviewer.md'), '# Reviewer\n\nReviews code changes for quality.\n')
-    writeFileSync(join(agentDir, 'deployer.md'), '# Deployer\n\nDeploys to staging environment.\n')
-
-    try {
-      const mockSession = createMockSession()
-      mockSession.cwd = tmpDir
-
-      server = new WsServer({
-        port: 0,
-        apiToken: TOKEN,
-        cliSession: mockSession,
-        authRequired: true,
-      })
-      const port = await startServerAndGetPort(server)
-      const { ws, messages } = await createClient(port, false)
-      send(ws, { type: 'auth', token: TOKEN })
-      await waitForMessage(messages, 'auth_ok', 2000)
-      messages.length = 0
-
-      send(ws, { type: 'list_agents' })
-      const result = await waitForMessage(messages, 'agent_list', 2000)
-
-      assert.ok(result, 'Should receive agent_list')
-      assert.ok(Array.isArray(result.agents))
-      assert.ok(result.agents.length >= 2, 'Should find at least 2 agents')
-
-      const deployer = result.agents.find(a => a.name === 'deployer')
-      assert.ok(deployer, 'Should include deployer agent')
-      assert.equal(deployer.source, 'project')
-      assert.ok(deployer.description.length > 0)
-
-      ws.close()
-    } finally {
-      rmSync(tmpDir, { recursive: true, force: true })
-    }
-  })
-
-  it('returns empty array when no agents exist', async () => {
-    const { mkdirSync, rmSync } = await import('fs')
-    const { join } = await import('path')
-    const { tmpdir } = await import('os')
-
-    const tmpDir = join(tmpdir(), `chroxy-test-agents-empty-${Date.now()}`)
-    mkdirSync(tmpDir, { recursive: true })
-
-    try {
-      const mockSession = createMockSession()
-      mockSession.cwd = tmpDir
-
-      server = new WsServer({
-        port: 0,
-        apiToken: TOKEN,
-        cliSession: mockSession,
-        authRequired: true,
-      })
-      const port = await startServerAndGetPort(server)
-      const { ws, messages } = await createClient(port, false)
-      send(ws, { type: 'auth', token: TOKEN })
-      await waitForMessage(messages, 'auth_ok', 2000)
-      messages.length = 0
-
-      send(ws, { type: 'list_agents' })
-      const result = await waitForMessage(messages, 'agent_list', 2000)
-
-      assert.ok(result, 'Should receive agent_list')
-      assert.ok(Array.isArray(result.agents))
-      // No project agents should exist (temp dir has no .claude/agents/)
-      // User agents from ~/.claude/agents/ may be present on the dev machine
-      const projectAgents = result.agents.filter(a => a.source === 'project')
-      assert.equal(projectAgents.length, 0, 'Should have no project agents from empty temp dir')
-
-      ws.close()
-    } finally {
-      rmSync(tmpDir, { recursive: true, force: true })
-    }
-  })
-
-  it('includes sessionId in multi-session mode', async () => {
-    const { mkdirSync, writeFileSync, rmSync } = await import('fs')
-    const { join } = await import('path')
-    const { tmpdir } = await import('os')
-
-    const tmpDir = join(tmpdir(), `chroxy-test-agents-ms-${Date.now()}`)
-    const agentDir = join(tmpDir, '.claude', 'agents')
-    mkdirSync(agentDir, { recursive: true })
-    writeFileSync(join(agentDir, 'helper.md'), '# Helper\n\nHelps with tasks.\n')
-
-    try {
-      const manager = new EventEmitter()
-      const mockSession = createMockSession()
-      mockSession.cwd = tmpDir
-
-      const sessionsMap = new Map()
-      sessionsMap.set('sess-1', { session: mockSession, name: 'Test', cwd: tmpDir, type: 'cli', isBusy: false })
-      manager.getSession = (id) => sessionsMap.get(id)
-      manager.listSessions = () => [{ id: 'sess-1', name: 'Test', cwd: tmpDir, type: 'cli', isBusy: false }]
-      manager.getHistory = () => []
-      manager.recordUserInput = () => {}
-      manager.getFullHistoryAsync = async () => []
-      manager.isBudgetPaused = () => false
-      Object.defineProperty(manager, 'firstSessionId', { get: () => 'sess-1' })
-
-      server = new WsServer({
-        port: 0,
-        apiToken: TOKEN,
-        sessionManager: manager,
-        authRequired: true,
-      })
-      const port = await startServerAndGetPort(server)
-      const { ws, messages } = await createClient(port, false)
-      send(ws, { type: 'auth', token: TOKEN })
-      await waitForMessage(messages, 'auth_ok', 2000)
-      messages.length = 0
-
-      send(ws, { type: 'list_agents' })
-      const result = await waitForMessage(messages, 'agent_list', 2000)
-
-      assert.ok(result, 'Should receive agent_list in multi-session mode')
-      assert.equal(result.sessionId, 'sess-1', 'agent_list should include sessionId in multi-session mode')
-      const helper = result.agents.find(a => a.name === 'helper')
-      assert.ok(helper, 'Should include helper agent')
-
-      ws.close()
-    } finally {
-      rmSync(tmpDir, { recursive: true, force: true })
-    }
-  })
-
-  it('omits sessionId in single-session CLI mode', async () => {
-    const { mkdirSync, rmSync } = await import('fs')
-    const { join } = await import('path')
-    const { tmpdir } = await import('os')
-
-    const tmpDir = join(tmpdir(), `chroxy-test-agents-cli-${Date.now()}`)
-    mkdirSync(tmpDir, { recursive: true })
-
-    try {
-      const mockSession = createMockSession()
-      mockSession.cwd = tmpDir
-
-      server = new WsServer({
-        port: 0,
-        apiToken: TOKEN,
-        cliSession: mockSession,
-        authRequired: true,
-      })
-      const port = await startServerAndGetPort(server)
-      const { ws, messages } = await createClient(port, false)
-      send(ws, { type: 'auth', token: TOKEN })
-      await waitForMessage(messages, 'auth_ok', 2000)
-      messages.length = 0
-
-      send(ws, { type: 'list_agents' })
-      const result = await waitForMessage(messages, 'agent_list', 2000)
-
-      assert.ok(result, 'Should receive agent_list')
-      assert.equal(result.sessionId, undefined, 'agent_list should NOT include sessionId in single-session mode')
-
-      ws.close()
-    } finally {
-      rmSync(tmpDir, { recursive: true, force: true })
-    }
-  })
-})
-
-describe('permission/question routing to originating session', () => {
-  const TOKEN = 'test-token'
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  function createTwoSessionManager() {
-    const manager = new EventEmitter()
-    const sessionsMap = new Map()
-
-    const sessionA = createMockSession()
-    sessionA.cwd = '/tmp/a'
-    sessionA.respondToPermission = () => {}
-    sessionA.respondToQuestion = () => {}
-    sessionsMap.set('sess-a', {
-      session: sessionA,
-      name: 'Session A',
-      cwd: '/tmp/a',
-      type: 'cli',
-      isBusy: false,
-    })
-
-    const sessionB = createMockSession()
-    sessionB.cwd = '/tmp/b'
-    sessionB.respondToPermission = () => {}
-    sessionB.respondToQuestion = () => {}
-    sessionsMap.set('sess-b', {
-      session: sessionB,
-      name: 'Session B',
-      cwd: '/tmp/b',
-      type: 'cli',
-      isBusy: false,
-    })
-
-    manager.getSession = (id) => sessionsMap.get(id)
-    manager.listSessions = () => {
-      const list = []
-      for (const [id, entry] of sessionsMap) {
-        list.push({ sessionId: id, name: entry.name, cwd: entry.cwd, type: entry.type, isBusy: entry.isBusy })
-      }
-      return list
-    }
-    manager.getHistory = () => []
-    manager.recordUserInput = () => {}
-    manager.touchActivity = () => {}
-    manager.getFullHistoryAsync = async () => []
-    manager.isBudgetPaused = () => false
-    manager.getSessionContext = async () => null
-    Object.defineProperty(manager, 'firstSessionId', {
-      get: () => 'sess-a'
-    })
-
-    return { manager, sessionsMap }
-  }
-
-  it('routes permission_response to the originating session, not activeSessionId', async () => {
-    const { manager, sessionsMap } = createTwoSessionManager()
-
-    let sessionAGotPermission = false
-    let sessionBGotPermission = false
-    sessionsMap.get('sess-a').session.respondToPermission = () => { sessionAGotPermission = true }
-    sessionsMap.get('sess-b').session.respondToPermission = () => { sessionBGotPermission = true }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      sessionManager: manager,
-      defaultSessionId: 'sess-a',
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, true)
-
-    // Simulate permission request from Session A (populates _permissionSessionMap)
-    server._permissionSessionMap.set('perm-routing-1', 'sess-a')
-
-    // Switch client to Session B
-    send(ws, { type: 'switch_session', sessionId: 'sess-b' })
-    await waitForMessage(messages, 'session_switched', 2000)
-
-    // Respond to the permission — should route to Session A despite active being B
-    send(ws, { type: 'permission_response', requestId: 'perm-routing-1', decision: 'allow' })
-    await waitFor(() => sessionAGotPermission, { label: 'sessionA permission routed' })
-
-    assert.equal(sessionAGotPermission, true, 'Session A should receive the permission response')
-    assert.equal(sessionBGotPermission, false, 'Session B should NOT receive the permission response')
-
-    ws.close()
-  })
-
-  it('routes user_question_response to the originating session, not activeSessionId', async () => {
-    const { manager, sessionsMap } = createTwoSessionManager()
-
-    let sessionAGotQuestion = false
-    let sessionBGotQuestion = false
-    sessionsMap.get('sess-a').session.respondToQuestion = () => { sessionAGotQuestion = true }
-    sessionsMap.get('sess-b').session.respondToQuestion = () => { sessionBGotQuestion = true }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      sessionManager: manager,
-      defaultSessionId: 'sess-a',
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, true)
-
-    // Simulate question from Session A (populates _questionSessionMap)
-    server._questionSessionMap.set('q-routing-1', 'sess-a')
-
-    // Switch client to Session B
-    send(ws, { type: 'switch_session', sessionId: 'sess-b' })
-    await waitForMessage(messages, 'session_switched', 2000)
-
-    // Respond to the question — should route to Session A
-    send(ws, { type: 'user_question_response', toolUseId: 'q-routing-1', answer: 'yes' })
-    await waitFor(() => sessionAGotQuestion, { label: 'sessionA question routed' })
-
-    assert.equal(sessionAGotQuestion, true, 'Session A should receive the question response')
-    assert.equal(sessionBGotQuestion, false, 'Session B should NOT receive the question response')
-
-    ws.close()
-  })
-
-  it('cleans up routing maps after permission response', async () => {
-    const { manager, sessionsMap } = createTwoSessionManager()
-    sessionsMap.get('sess-a').session.respondToPermission = () => {}
-
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      sessionManager: manager,
-      defaultSessionId: 'sess-a',
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws } = await createClient(port, true)
-
-    // Populate routing map
-    server._permissionSessionMap.set('perm-cleanup-1', 'sess-a')
-    assert.equal(server._permissionSessionMap.size, 1)
-
-    // Respond
-    send(ws, { type: 'permission_response', requestId: 'perm-cleanup-1', decision: 'allow' })
-    await waitFor(() => server._permissionSessionMap.size === 0, { label: 'permission map cleaned up' })
-
-    // Map entry should be deleted
-    assert.equal(server._permissionSessionMap.size, 0, 'Permission routing map should be cleaned up')
-
-    ws.close()
-  })
-
-  it('cleans up routing maps after question response', async () => {
-    const { manager, sessionsMap } = createTwoSessionManager()
-    sessionsMap.get('sess-a').session.respondToQuestion = () => {}
-
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      sessionManager: manager,
-      defaultSessionId: 'sess-a',
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws } = await createClient(port, true)
-
-    // Populate routing map
-    server._questionSessionMap.set('q-cleanup-1', 'sess-a')
-    assert.equal(server._questionSessionMap.size, 1)
-
-    // Respond
-    send(ws, { type: 'user_question_response', toolUseId: 'q-cleanup-1', answer: 'yes' })
-    await waitFor(() => server._questionSessionMap.size === 0, { label: 'question map cleaned up' })
-
-    // Map entry should be deleted
-    assert.equal(server._questionSessionMap.size, 0, 'Question routing map should be cleaned up')
-
-    ws.close()
-  })
-
-  it('falls back to activeSessionId for unknown permission requestId', async () => {
-    const { manager, sessionsMap } = createTwoSessionManager()
-
-    let sessionBGotPermission = false
-    sessionsMap.get('sess-b').session.respondToPermission = () => { sessionBGotPermission = true }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      sessionManager: manager,
-      defaultSessionId: 'sess-a',
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, true)
-
-    // Switch to Session B
-    send(ws, { type: 'switch_session', sessionId: 'sess-b' })
-    await waitForMessage(messages, 'session_switched', 2000)
-
-    // Send permission response with unknown requestId — no entry in routing map
-    // Should fall back to activeSessionId (sess-b)
-    send(ws, { type: 'permission_response', requestId: 'unknown-id', decision: 'allow' })
-    await waitFor(() => sessionBGotPermission, { label: 'sessionB fallback permission' })
-
-    assert.equal(sessionBGotPermission, true, 'Should fall back to activeSessionId when requestId not in routing map')
-
-    ws.close()
-  })
-
-  it('falls back to activeSessionId for unknown question toolUseId', async () => {
-    const { manager, sessionsMap } = createTwoSessionManager()
-
-    let sessionBGotQuestion = false
-    sessionsMap.get('sess-b').session.respondToQuestion = () => { sessionBGotQuestion = true }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: TOKEN,
-      sessionManager: manager,
-      defaultSessionId: 'sess-a',
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, true)
-
-    // Switch to Session B
-    send(ws, { type: 'switch_session', sessionId: 'sess-b' })
-    await waitForMessage(messages, 'session_switched', 2000)
-
-    // Send question response with unknown toolUseId — no entry in routing map
-    // Should fall back to activeSessionId (sess-b)
-    send(ws, { type: 'user_question_response', toolUseId: 'unknown-tool-use-id', answer: 'yes' })
-    await waitFor(() => sessionBGotQuestion, { label: 'sessionB fallback question' })
-
-    assert.equal(sessionBGotQuestion, true, 'Should fall back to activeSessionId when toolUseId not in routing map')
-
-    ws.close()
-  })
-})
-
-describe('request_session_context error paths', () => {
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('returns session_error when no active session', async () => {
-    // Manager with no sessions — no activeSessionId can be set
-    const { manager: mockManager } = createMockSessionManager([])
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, true)
-
-    send(ws, { type: 'request_session_context' })
-    const err = await waitForMessage(messages, 'session_error', 2000)
-    assert.equal(err.message, 'No active session')
-
-    ws.close()
-  })
-
-  it('returns session_error when session not found', async () => {
-    const { manager: mockManager } = createMockSessionManager([
-      { id: 'sess-1', name: 'Test', cwd: '/tmp' }
-    ])
-    mockManager.getSessionContext = async () => null
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, true)
-
-    // Request context for a non-existent session — getSessionContext returns null
-    send(ws, { type: 'request_session_context', sessionId: 'nonexistent' })
-    const err = await waitForMessage(messages, 'session_error', 2000)
-    assert.ok(err.message.includes('Session not found: nonexistent'))
-
-    ws.close()
-  })
-
-  it('returns session_error when getSessionContext throws', async () => {
-    const { manager: mockManager } = createMockSessionManager([
-      { id: 'sess-1', name: 'Test', cwd: '/tmp' }
-    ])
-    mockManager.getSessionContext = async () => { throw new Error('git not found') }
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, true)
-
-    // Request context for a valid session but getSessionContext throws
-    send(ws, { type: 'request_session_context', sessionId: 'sess-1' })
-    const err = await waitForMessage(messages, 'session_error', 2000)
-    assert.ok(err.message.includes('Failed to read session context'))
-    assert.ok(err.message.includes('git not found'))
-
-    ws.close()
-  })
-})
-
-describe('POST /permission-response HTTP endpoint', () => {
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('resolves a pending legacy permission via HTTP', async () => {
-    const mockPty = new EventEmitter()
-    mockPty.write = () => {}
-    mockPty.resize = () => {}
-    const mockParser = new EventEmitter()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      ptyManager: mockPty,
-      outputParser: mockParser,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, true)
-
-    // Create a pending permission via POST /permission
-    const permPromise = fetch(`http://127.0.0.1:${port}/permission`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ tool_name: 'Write', tool_input: { file_path: '/tmp/test.js' } }),
-    })
-
-    try {
-      const permReq = await waitForMessage(messages, 'permission_request', 2000)
-
-      // Resolve via POST /permission-response
-      const response = await fetch(`http://127.0.0.1:${port}/permission-response`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ requestId: permReq.requestId, decision: 'allow' }),
-      })
-      assert.equal(response.status, 200)
-      const data = await response.json()
-      assert.deepEqual(data, { ok: true })
-
-      // The original /permission response should also complete
-      const permRes = await permPromise
-      assert.equal(permRes.status, 200)
-      const permData = await permRes.json()
-      assert.equal(permData.decision, 'allow')
-    } finally {
-      ws.close()
-      await permPromise.catch(() => {})
-    }
-  })
-
-  it('rejects unauthenticated requests when authRequired: true', async () => {
-    const mockPty = new EventEmitter()
-    mockPty.write = () => {}
-    mockPty.resize = () => {}
-    const mockParser = new EventEmitter()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      ptyManager: mockPty,
-      outputParser: mockParser,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const response = await fetch(`http://127.0.0.1:${port}/permission-response`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ requestId: 'perm-1', decision: 'allow' }),
-    })
-    assert.equal(response.status, 403)
-    const data = await response.json()
-    assert.equal(data.error, 'unauthorized')
-  })
-
-  it('authenticates with Bearer token when authRequired: true', async () => {
-    const mockPty = new EventEmitter()
-    mockPty.write = () => {}
-    mockPty.resize = () => {}
-    const mockParser = new EventEmitter()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      ptyManager: mockPty,
-      outputParser: mockParser,
-      authRequired: true,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, false)
-    // Authenticate the WS client
-    send(ws, { type: 'auth', token: 'test-token' })
-    await waitForMessage(messages, 'auth_ok', 2000)
-
-    // Create a pending permission
-    const permPromise = fetch(`http://127.0.0.1:${port}/permission`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'Authorization': 'Bearer test-token',
-      },
-      body: JSON.stringify({ tool_name: 'Bash', tool_input: { command: 'ls' } }),
-    })
-
-    try {
-      const permReq = await waitForMessage(messages, 'permission_request', 2000)
-
-      // Resolve via authenticated POST /permission-response
-      const response = await fetch(`http://127.0.0.1:${port}/permission-response`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          'Authorization': 'Bearer test-token',
-        },
-        body: JSON.stringify({ requestId: permReq.requestId, decision: 'allow' }),
-      })
-      assert.equal(response.status, 200)
-      const data = await response.json()
-      assert.deepEqual(data, { ok: true })
-    } finally {
-      ws.close()
-      await permPromise.catch(() => {})
-    }
-  })
-
-  it('returns 404 for unknown requestId', async () => {
-    const mockPty = new EventEmitter()
-    mockPty.write = () => {}
-    mockPty.resize = () => {}
-    const mockParser = new EventEmitter()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      ptyManager: mockPty,
-      outputParser: mockParser,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const response = await fetch(`http://127.0.0.1:${port}/permission-response`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ requestId: 'nonexistent-123', decision: 'allow' }),
-    })
-    assert.equal(response.status, 404)
-    const data = await response.json()
-    assert.equal(data.error, 'unknown or expired requestId')
-  })
-
-  it('returns 400 for invalid decision', async () => {
-    const mockPty = new EventEmitter()
-    mockPty.write = () => {}
-    mockPty.resize = () => {}
-    const mockParser = new EventEmitter()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      ptyManager: mockPty,
-      outputParser: mockParser,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const response = await fetch(`http://127.0.0.1:${port}/permission-response`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ requestId: 'perm-1', decision: 'maybe' }),
-    })
-    assert.equal(response.status, 400)
-    const data = await response.json()
-    assert.ok(data.error.includes('invalid decision'))
-  })
-
-  it('returns 400 for missing requestId', async () => {
-    const mockPty = new EventEmitter()
-    mockPty.write = () => {}
-    mockPty.resize = () => {}
-    const mockParser = new EventEmitter()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      ptyManager: mockPty,
-      outputParser: mockParser,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const response = await fetch(`http://127.0.0.1:${port}/permission-response`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ decision: 'allow' }),
-    })
-    assert.equal(response.status, 400)
-    const data = await response.json()
-    assert.equal(data.error, 'missing requestId')
-  })
-
-  it('returns 400 for invalid JSON body', async () => {
-    const mockPty = new EventEmitter()
-    mockPty.write = () => {}
-    mockPty.resize = () => {}
-    const mockParser = new EventEmitter()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      ptyManager: mockPty,
-      outputParser: mockParser,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const response = await fetch(`http://127.0.0.1:${port}/permission-response`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: 'not json',
-    })
-    assert.equal(response.status, 400)
-    const data = await response.json()
-    assert.equal(data.error, 'invalid JSON')
-  })
-
-  it('returns 404 on duplicate response (already resolved)', async () => {
-    const mockPty = new EventEmitter()
-    mockPty.write = () => {}
-    mockPty.resize = () => {}
-    const mockParser = new EventEmitter()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      ptyManager: mockPty,
-      outputParser: mockParser,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, true)
-
-    // Create a pending permission
-    const permPromise = fetch(`http://127.0.0.1:${port}/permission`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ tool_name: 'Read', tool_input: { file_path: '/tmp/file.txt' } }),
-    })
-
-    try {
-      const permReq = await waitForMessage(messages, 'permission_request', 2000)
-
-      // First response succeeds
-      const first = await fetch(`http://127.0.0.1:${port}/permission-response`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ requestId: permReq.requestId, decision: 'allow' }),
-      })
-      assert.equal(first.status, 200)
-
-      // Second response returns 404 (already resolved)
-      const second = await fetch(`http://127.0.0.1:${port}/permission-response`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ requestId: permReq.requestId, decision: 'deny' }),
-      })
-      assert.equal(second.status, 404)
-    } finally {
-      ws.close()
-      await permPromise.catch(() => {})
-    }
-  })
-
-  it('resolves SDK session permission via HTTP', async () => {
-    const manager = new EventEmitter()
-    const sessionsMap = new Map()
-    const mockSession = createMockSession()
-    mockSession.cwd = '/tmp/project'
-    let resolvedWith = null
-    mockSession.respondToPermission = (reqId, decision) => { resolvedWith = { reqId, decision } }
-    sessionsMap.set('sess-1', {
-      session: mockSession,
-      name: 'Session 1',
-      cwd: '/tmp/project',
-      type: 'cli',
-      isBusy: false,
-    })
-    manager.getSession = (id) => sessionsMap.get(id)
-    manager.listSessions = () => [{ id: 'sess-1', name: 'Session 1', cwd: '/tmp/project', type: 'cli', isBusy: false }]
-    manager.getHistory = () => []
-    manager.recordUserInput = () => {}
-    manager.getFullHistoryAsync = async () => []
-    manager.isBudgetPaused = () => false
-    manager.getSessionContext = async () => null
-    Object.defineProperty(manager, 'firstSessionId', {
-      get: () => 'sess-1'
-    })
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: manager,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, true)
-
-    // Wait for the client to be fully set up (session_switched, etc.)
-    await waitForMessage(messages, 'session_switched', 2000)
-
-    // Simulate SDK session emitting a permission_request
-    const requestId = 'sdk-perm-http-1'
-    manager.emit('session_event', {
-      sessionId: 'sess-1',
-      event: 'permission_request',
-      data: {
-        requestId,
-        tool: 'Bash',
-        description: 'ls',
-        input: { command: 'ls' },
-        remainingMs: 300000,
-      },
-    })
-
-    await waitForMessage(messages, 'permission_request', 2000)
-
-    // Resolve via HTTP
-    const response = await fetch(`http://127.0.0.1:${port}/permission-response`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ requestId, decision: 'allow' }),
-    })
-    assert.equal(response.status, 200)
-    const data = await response.json()
-    assert.deepEqual(data, { ok: true })
-    assert.deepEqual(resolvedWith, { reqId: requestId, decision: 'allow' })
-
-    ws.close()
-  })
-})
-
-// ---------------------------------------------------------------------------
-// File browser symlink traversal tests (#690)
-// ---------------------------------------------------------------------------
-describe('file browser symlink security', () => {
-  let server
-  let tempDir    // main CWD
-  let outsideDir // directory outside CWD that symlinks target
-
-  beforeEach(() => {
-    // Create temp directories
-    tempDir = mkdtempSync(join(tmpdir(), 'chroxy-test-cwd-'))
-    outsideDir = mkdtempSync(join(tmpdir(), 'chroxy-test-outside-'))
-
-    // Create structure inside CWD:
-    //   tempDir/
-    //     subdir/
-    //       file.txt
-    //     internal-link -> subdir/     (symlink within CWD — should work)
-    //     escape-link -> outsideDir/   (symlink outside CWD — should be blocked)
-    //     escape-file -> outsideDir/secret.txt (file symlink outside CWD — should be blocked)
-    mkdirSync(join(tempDir, 'subdir'))
-    writeFileSync(join(tempDir, 'subdir', 'file.txt'), 'inside content')
-    writeFileSync(join(outsideDir, 'secret.txt'), 'outside secret')
-    mkdirSync(join(outsideDir, 'hidden-dir'))
-    writeFileSync(join(outsideDir, 'hidden-dir', 'data.txt'), 'hidden data')
-
-    symlinkSync(join(tempDir, 'subdir'), join(tempDir, 'internal-link'))
-    symlinkSync(outsideDir, join(tempDir, 'escape-link'))
-    symlinkSync(join(outsideDir, 'secret.txt'), join(tempDir, 'escape-file'))
-  })
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-    rmSync(tempDir, { recursive: true, force: true })
-    rmSync(outsideDir, { recursive: true, force: true })
-  })
-
-  /** Spin up a WsServer with cwd set to tempDir and return a connected client. */
-  async function createFileBrowserTestServer() {
-    const mockSession = createMockSession()
-    mockSession.cwd = tempDir
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, true)
-    return { ws, messages }
-  }
-
-  it('browse_files: rejects symlink directory pointing outside CWD', async () => {
-    const { ws, messages } = await createFileBrowserTestServer()
-
-    send(ws, { type: 'browse_files', path: 'escape-link' })
-    const listing = await waitForMessage(messages, 'file_listing', 2000)
-
-    assert.ok(listing.error, 'Should return an error for symlink outside CWD')
-    assert.match(listing.error, /access denied/i)
-    assert.deepEqual(listing.entries, [])
-
-    ws.close()
-  })
-
-  it('browse_files: allows symlink directory pointing within CWD', async () => {
-    const { ws, messages } = await createFileBrowserTestServer()
-
-    send(ws, { type: 'browse_files', path: 'internal-link' })
-    const listing = await waitForMessage(messages, 'file_listing', 2000)
-
-    assert.equal(listing.error, null, 'Should not return an error for symlink within CWD')
-    assert.ok(listing.entries.length > 0, 'Should return entries')
-    assert.ok(listing.entries.some(e => e.name === 'file.txt'), 'Should list file.txt inside symlinked dir')
-
-    ws.close()
-  })
-
-  it('browse_files: rejects ../../../ path traversal', async () => {
-    const { ws, messages } = await createFileBrowserTestServer()
-
-    send(ws, { type: 'browse_files', path: '../../../etc' })
-    const listing = await waitForMessage(messages, 'file_listing', 2000)
-
-    assert.ok(listing.error, 'Should return an error for path traversal')
-    assert.match(listing.error, /access denied/i)
-    assert.deepEqual(listing.entries, [])
-
-    ws.close()
-  })
-
-  it('read_file: rejects symlink file pointing outside CWD', async () => {
-    const { ws, messages } = await createFileBrowserTestServer()
-
-    send(ws, { type: 'read_file', path: 'escape-file' })
-    const content = await waitForMessage(messages, 'file_content', 2000)
-
-    assert.ok(content.error, 'Should return an error for symlink file outside CWD')
-    assert.match(content.error, /access denied/i)
-    assert.equal(content.content, null)
-
-    ws.close()
-  })
-
-  it('read_file: allows reading file through symlink within CWD', async () => {
-    const { ws, messages } = await createFileBrowserTestServer()
-
-    send(ws, { type: 'read_file', path: 'internal-link/file.txt' })
-    const content = await waitForMessage(messages, 'file_content', 2000)
-
-    assert.equal(content.error, null, 'Should not return error for symlink within CWD')
-    assert.equal(content.content, 'inside content')
-
-    ws.close()
-  })
-
-  it('read_file: rejects ../../../etc/passwd traversal', async () => {
-    const { ws, messages } = await createFileBrowserTestServer()
-
-    send(ws, { type: 'read_file', path: '../../../etc/passwd' })
-    const content = await waitForMessage(messages, 'file_content', 2000)
-
-    assert.ok(content.error, 'Should return an error for path traversal')
-    assert.match(content.error, /access denied/i)
-    assert.equal(content.content, null)
-
-    ws.close()
-  })
-
-  it('read_file: rejects null bytes in path', async () => {
-    const { ws, messages } = await createFileBrowserTestServer()
-
-    send(ws, { type: 'read_file', path: 'subdir/file.txt\x00.jpg' })
-    const content = await waitForMessage(messages, 'file_content', 2000)
-
-    // Should error — either access denied or file not found, but NOT return content
-    assert.ok(content.error, 'Should return an error for null bytes in path')
-
-    ws.close()
-  })
-
-  it('browse_files: rejects symlink chain escaping CWD', async () => {
-    // Create a chain: tempDir/chain-link -> outsideDir/hidden-dir
-    symlinkSync(join(outsideDir, 'hidden-dir'), join(tempDir, 'chain-link'))
-
-    const { ws, messages } = await createFileBrowserTestServer()
-
-    send(ws, { type: 'browse_files', path: 'chain-link' })
-    const listing = await waitForMessage(messages, 'file_listing', 2000)
-
-    assert.ok(listing.error, 'Should return an error for symlink chain outside CWD')
-    assert.match(listing.error, /access denied/i)
-
-    ws.close()
-  })
-})
-
-// ---------------------------------------------------------------------------
-// Encryption key exchange timeout/rejection tests (#691)
-// Uses _WsServer (raw, encryption enabled) instead of the WsServer wrapper
-// ---------------------------------------------------------------------------
 describe('encryption key exchange enforcement', () => {
   let server
 
@@ -3711,377 +968,6 @@ describe('encryption key exchange enforcement', () => {
     assert.equal(authOk.encryption, 'disabled', 'Should indicate encryption is disabled')
 
     ws.close()
-  })
-})
-
-describe('Reconnect permission recovery (_resendPendingPermissions)', () => {
-  let server
-
-  /**
-   * Create a WsServer with a mock sessionManager that exposes _sessions Map.
-   * Each session has _pendingPermissions and _lastPermissionData maps
-   * to simulate SDK-mode permission state.
-   */
-  function createPermMockSessionManager(sessions = {}) {
-    const _sessions = new Map()
-    for (const [id, { pending, lastData }] of Object.entries(sessions)) {
-      const session = new EventEmitter()
-      session._pendingPermissions = new Map(Object.entries(pending || {}))
-      session._lastPermissionData = new Map(Object.entries(lastData || {}))
-      _sessions.set(id, { session })
-    }
-    return { _sessions }
-  }
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('re-sends pending permission on reconnect with correct fields', () => {
-    const now = Date.now()
-    const permData = {
-      requestId: 'perm-1',
-      tool: 'Bash',
-      description: 'ls -la',
-      input: { command: 'ls -la' },
-      remainingMs: 300_000,
-      createdAt: now - 1000, // 1 second ago
-    }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      authRequired: false,
-    })
-    server.sessionManager = createPermMockSessionManager({
-      'session-1': {
-        pending: { 'perm-1': true },
-        lastData: { 'perm-1': permData },
-      },
-    })
-
-    const sent = []
-    server._send = (ws, msg) => sent.push(msg)
-
-    server._resendPendingPermissions({})
-
-    assert.equal(sent.length, 1, 'Should send exactly one permission')
-    assert.equal(sent[0].type, 'permission_request')
-    assert.equal(sent[0].requestId, 'perm-1')
-    assert.equal(sent[0].tool, 'Bash')
-    assert.equal(sent[0].sessionId, 'session-1')
-    assert.ok(sent[0].remainingMs > 0 && sent[0].remainingMs <= 300_000,
-      'remainingMs should be positive and <= 300s')
-    // createdAt is internal server state — must NOT leak to client protocol
-    assert.equal(sent[0].createdAt, undefined, 'createdAt should not be sent to client')
-  })
-
-  it('adjusts remainingMs for elapsed time', () => {
-    const twoMinAgo = Date.now() - 120_000
-    const permData = {
-      requestId: 'perm-2',
-      tool: 'Edit',
-      description: 'edit file',
-      input: {},
-      remainingMs: 300_000,
-      createdAt: twoMinAgo,
-    }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      authRequired: false,
-    })
-    server.sessionManager = createPermMockSessionManager({
-      'session-1': {
-        pending: { 'perm-2': true },
-        lastData: { 'perm-2': permData },
-      },
-    })
-
-    const sent = []
-    server._send = (ws, msg) => sent.push(msg)
-
-    server._resendPendingPermissions({})
-
-    assert.equal(sent.length, 1)
-    // Should be approximately 180s remaining (300 - 120), allow 5s tolerance
-    assert.ok(sent[0].remainingMs >= 175_000 && sent[0].remainingMs <= 185_000,
-      `Expected ~180s remaining, got ${sent[0].remainingMs}ms`)
-  })
-
-  it('skips expired permissions (createdAt > 5 min ago)', () => {
-    const sixMinAgo = Date.now() - 360_000
-    const permData = {
-      requestId: 'perm-3',
-      tool: 'Bash',
-      description: 'expired command',
-      input: {},
-      remainingMs: 300_000,
-      createdAt: sixMinAgo,
-    }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      authRequired: false,
-    })
-    server.sessionManager = createPermMockSessionManager({
-      'session-1': {
-        pending: { 'perm-3': true },
-        lastData: { 'perm-3': permData },
-      },
-    })
-
-    const sent = []
-    server._send = (ws, msg) => sent.push(msg)
-
-    server._resendPendingPermissions({})
-
-    assert.equal(sent.length, 0, 'Should not re-send expired permissions')
-  })
-
-  it('skips resolved permissions (in lastData but not pendingPermissions)', () => {
-    const permData = {
-      requestId: 'perm-4',
-      tool: 'Read',
-      description: 'read file',
-      input: {},
-      remainingMs: 300_000,
-      createdAt: Date.now() - 10_000,
-    }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      authRequired: false,
-    })
-    // Permission is in lastData but NOT in pendingPermissions (already resolved)
-    server.sessionManager = createPermMockSessionManager({
-      'session-1': {
-        pending: {}, // empty — permission was resolved
-        lastData: { 'perm-4': permData },
-      },
-    })
-
-    const sent = []
-    server._send = (ws, msg) => sent.push(msg)
-
-    server._resendPendingPermissions({})
-
-    assert.equal(sent.length, 0, 'Should not re-send resolved permissions')
-  })
-
-  it('handles multiple sessions with mixed pending state', () => {
-    const now = Date.now()
-    const activePerm = {
-      requestId: 'perm-5',
-      tool: 'Bash',
-      description: 'active perm',
-      input: {},
-      remainingMs: 300_000,
-      createdAt: now - 30_000, // 30s ago
-    }
-    const resolvedPerm = {
-      requestId: 'perm-6',
-      tool: 'Edit',
-      description: 'resolved perm',
-      input: {},
-      remainingMs: 300_000,
-      createdAt: now - 60_000,
-    }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      authRequired: false,
-    })
-    server.sessionManager = createPermMockSessionManager({
-      'session-a': {
-        pending: { 'perm-5': true },
-        lastData: { 'perm-5': activePerm },
-      },
-      'session-b': {
-        pending: {}, // perm-6 was resolved
-        lastData: { 'perm-6': resolvedPerm },
-      },
-    })
-
-    const sent = []
-    server._send = (ws, msg) => sent.push(msg)
-
-    server._resendPendingPermissions({})
-
-    assert.equal(sent.length, 1, 'Should only send the active permission')
-    assert.equal(sent[0].requestId, 'perm-5')
-    assert.equal(sent[0].sessionId, 'session-a')
-  })
-
-  it('re-sends legacy HTTP-held permissions with adjusted remainingMs', () => {
-    const now = Date.now()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      authRequired: false,
-    })
-    // No session manager — use legacy _pendingPermissions
-    server.sessionManager = null
-    server._pendingPermissions.set('perm-legacy', {
-      resolve: () => {},
-      timer: null,
-      data: {
-        requestId: 'perm-legacy',
-        tool: 'Bash',
-        description: 'legacy cmd',
-        input: {},
-        remainingMs: 300_000,
-        createdAt: now - 60_000, // 1 minute ago
-      },
-    })
-
-    const sent = []
-    server._send = (ws, msg) => sent.push(msg)
-
-    server._resendPendingPermissions({})
-
-    assert.equal(sent.length, 1)
-    assert.equal(sent[0].requestId, 'perm-legacy')
-    // Should be approximately 240s remaining (300 - 60), allow 5s tolerance
-    assert.ok(sent[0].remainingMs >= 235_000 && sent[0].remainingMs <= 245_000,
-      `Expected ~240s remaining, got ${sent[0].remainingMs}ms`)
-    // createdAt must not leak to client
-    assert.equal(sent[0].createdAt, undefined, 'createdAt should not be sent to client')
-  })
-
-  it('re-populates _permissionSessionMap so responses route to correct session after reconnect', () => {
-    const now = Date.now()
-    const permData = {
-      requestId: 'perm-route-1',
-      tool: 'Bash',
-      description: 'route test',
-      input: { command: 'echo hi' },
-      remainingMs: 300_000,
-      createdAt: now - 5000,
-    }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      authRequired: false,
-    })
-    server.sessionManager = createPermMockSessionManager({
-      'session-x': {
-        pending: { 'perm-route-1': true },
-        lastData: { 'perm-route-1': permData },
-      },
-    })
-
-    // Clear any pre-existing routing entries
-    server._permissionSessionMap.clear()
-
-    const sent = []
-    server._send = (ws, msg) => sent.push(msg)
-
-    server._resendPendingPermissions({})
-
-    // Verify the permission was re-sent
-    assert.equal(sent.length, 1, 'Should re-send the permission')
-    assert.equal(sent[0].requestId, 'perm-route-1')
-
-    // Verify the routing map was re-populated
-    assert.equal(server._permissionSessionMap.get('perm-route-1'), 'session-x',
-      'Should re-populate _permissionSessionMap so response routes to originating session')
-  })
-
-  it('re-populates routing map for multiple permissions across sessions', () => {
-    const now = Date.now()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      authRequired: false,
-    })
-    server.sessionManager = createPermMockSessionManager({
-      'session-a': {
-        pending: { 'perm-a1': true },
-        lastData: {
-          'perm-a1': {
-            requestId: 'perm-a1',
-            tool: 'Bash',
-            description: 'cmd a',
-            input: {},
-            remainingMs: 300_000,
-            createdAt: now - 10_000,
-          },
-        },
-      },
-      'session-b': {
-        pending: { 'perm-b1': true },
-        lastData: {
-          'perm-b1': {
-            requestId: 'perm-b1',
-            tool: 'Edit',
-            description: 'edit b',
-            input: {},
-            remainingMs: 300_000,
-            createdAt: now - 20_000,
-          },
-        },
-      },
-    })
-
-    server._permissionSessionMap.clear()
-
-    const sent = []
-    server._send = (ws, msg) => sent.push(msg)
-
-    server._resendPendingPermissions({})
-
-    assert.equal(sent.length, 2, 'Should re-send both permissions')
-    assert.equal(server._permissionSessionMap.get('perm-a1'), 'session-a',
-      'perm-a1 should route to session-a')
-    assert.equal(server._permissionSessionMap.get('perm-b1'), 'session-b',
-      'perm-b1 should route to session-b')
-  })
-
-  it('does not populate routing map for expired permissions', () => {
-    const sixMinAgo = Date.now() - 360_000
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      authRequired: false,
-    })
-    server.sessionManager = createPermMockSessionManager({
-      'session-1': {
-        pending: { 'perm-expired': true },
-        lastData: {
-          'perm-expired': {
-            requestId: 'perm-expired',
-            tool: 'Bash',
-            description: 'expired',
-            input: {},
-            remainingMs: 300_000,
-            createdAt: sixMinAgo,
-          },
-        },
-      },
-    })
-
-    server._permissionSessionMap.clear()
-
-    const sent = []
-    server._send = (ws, msg) => sent.push(msg)
-
-    server._resendPendingPermissions({})
-
-    assert.equal(sent.length, 0, 'Should not re-send expired permission')
-    assert.equal(server._permissionSessionMap.has('perm-expired'), false,
-      'Should not populate routing map for expired permissions')
   })
 })
 
@@ -4276,575 +1162,6 @@ describe('outbound message sequence numbers', () => {
 // ---------------------------------------------------------------------------
 // Diff handler tests (#607)
 // ---------------------------------------------------------------------------
-describe('get_diff handler', () => {
-  let server
-  let tempDir
-
-  beforeEach(() => {
-    tempDir = mkdtempSync(join(tmpdir(), 'chroxy-diff-test-'))
-    // Initialize a git repo in the temp directory
-    execFileSync(GIT, ['init'], { cwd: tempDir, stdio: 'pipe' })
-    execFileSync(GIT, ['config', 'user.email', 'test@test.com'], { cwd: tempDir, stdio: 'pipe' })
-    execFileSync(GIT, ['config', 'user.name', 'Test'], { cwd: tempDir, stdio: 'pipe' })
-    // Create an initial commit
-    writeFileSync(join(tempDir, 'file.txt'), 'initial content\n')
-    execFileSync(GIT, ['add', '.'], { cwd: tempDir, stdio: 'pipe' })
-    execFileSync(GIT, ['commit', '-m', 'initial'], { cwd: tempDir, stdio: 'pipe' })
-  })
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-    rmSync(tempDir, { recursive: true, force: true })
-  })
-
-  async function createDiffTestServer() {
-    const mockSession = createMockSession()
-    mockSession.cwd = tempDir
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, true)
-    return { ws, messages }
-  }
-
-  it('returns empty files array when no changes', async () => {
-    const { ws, messages } = await createDiffTestServer()
-
-    send(ws, { type: 'get_diff' })
-    const result = await waitForMessage(messages, 'diff_result', 5000)
-
-    assert.equal(result.error, null)
-    assert.deepEqual(result.files, [])
-
-    ws.close()
-  })
-
-  it('returns diff for modified file', async () => {
-    // Modify the file
-    writeFileSync(join(tempDir, 'file.txt'), 'modified content\n')
-
-    const { ws, messages } = await createDiffTestServer()
-
-    send(ws, { type: 'get_diff' })
-    const result = await waitForMessage(messages, 'diff_result', 5000)
-
-    assert.equal(result.error, null)
-    assert.equal(result.files.length, 1)
-    assert.equal(result.files[0].path, 'file.txt')
-    assert.equal(result.files[0].status, 'modified')
-    assert.ok(result.files[0].additions > 0 || result.files[0].deletions > 0,
-      'Should have additions or deletions')
-    assert.ok(result.files[0].hunks.length > 0, 'Should have hunks')
-
-    ws.close()
-  })
-
-  it('returns untracked new file with synthetic diff', async () => {
-    writeFileSync(join(tempDir, 'new-file.txt'), 'new content\n')
-
-    const { ws, messages } = await createDiffTestServer()
-
-    send(ws, { type: 'get_diff' })
-    const result = await waitForMessage(messages, 'diff_result', 5000)
-
-    assert.equal(result.error, null)
-    assert.equal(result.files.length, 1)
-    assert.equal(result.files[0].path, 'new-file.txt')
-    assert.equal(result.files[0].status, 'untracked')
-    assert.equal(result.files[0].additions, 1)
-    assert.equal(result.files[0].deletions, 0)
-    assert.equal(result.files[0].hunks.length, 1)
-    assert.equal(result.files[0].hunks[0].header, 'New untracked file')
-    assert.equal(result.files[0].hunks[0].lines[0].type, 'addition')
-    assert.equal(result.files[0].hunks[0].lines[0].content, 'new content')
-
-    ws.close()
-  })
-
-  it('shows untracked files alongside modified files', async () => {
-    writeFileSync(join(tempDir, 'file.txt'), 'modified content\n')
-    writeFileSync(join(tempDir, 'untracked.txt'), 'brand new\n')
-
-    const { ws, messages } = await createDiffTestServer()
-
-    send(ws, { type: 'get_diff' })
-    const result = await waitForMessage(messages, 'diff_result', 5000)
-
-    assert.equal(result.error, null)
-    assert.equal(result.files.length, 2)
-
-    const modified = result.files.find(f => f.path === 'file.txt')
-    const untracked = result.files.find(f => f.path === 'untracked.txt')
-    assert.ok(modified, 'Modified file should be present')
-    assert.ok(untracked, 'Untracked file should be present')
-    assert.equal(modified.status, 'modified')
-    assert.equal(untracked.status, 'untracked')
-
-    ws.close()
-  })
-
-  it('caps untracked files at 10', async () => {
-    for (let i = 0; i < 15; i++) {
-      writeFileSync(join(tempDir, `untracked-${String(i).padStart(2, '0')}.txt`), `content ${i}\n`)
-    }
-
-    const { ws, messages } = await createDiffTestServer()
-
-    send(ws, { type: 'get_diff' })
-    const result = await waitForMessage(messages, 'diff_result', 5000)
-
-    assert.equal(result.error, null)
-    const untrackedFiles = result.files.filter(f => f.status === 'untracked')
-    assert.equal(untrackedFiles.length, 10, 'Should cap at 10 untracked files')
-
-    ws.close()
-  })
-
-  it('shows placeholder for untracked files exceeding 50KB', async () => {
-    // Create a file just over 50KB
-    const bigContent = 'x'.repeat(51 * 1024) + '\n'
-    writeFileSync(join(tempDir, 'big-untracked.txt'), bigContent)
-
-    const { ws, messages } = await createDiffTestServer()
-
-    send(ws, { type: 'get_diff' })
-    const result = await waitForMessage(messages, 'diff_result', 5000)
-
-    assert.equal(result.error, null)
-    const bigFile = result.files.find(f => f.path === 'big-untracked.txt')
-    assert.ok(bigFile, 'Big untracked file should be present')
-    assert.equal(bigFile.status, 'untracked')
-    assert.equal(bigFile.additions, 0, 'Too-large file should have 0 additions')
-    assert.equal(bigFile.hunks.length, 1)
-    assert.equal(bigFile.hunks[0].lines.length, 1)
-    assert.equal(bigFile.hunks[0].lines[0].type, 'context')
-    assert.ok(bigFile.hunks[0].lines[0].content.includes('File too large to preview'), 'Should show size placeholder')
-
-    ws.close()
-  })
-
-  it('shows placeholder for binary untracked files', async () => {
-    // Create a binary file with realistic JPEG header bytes (invalid UTF-8 + null bytes)
-    const binaryContent = Buffer.from([
-      0xFF, 0xD8, 0xFF, 0xE0,                         // JPEG SOI + APP0 marker
-      0x00, 0x10, 0x4A, 0x46, 0x49, 0x46, 0x00,       // JFIF segment with nulls
-      0x01, 0x02, 0xFF, 0xDB, 0xFF, 0xC0, 0xFF, 0xDA, // typical JPEG markers
-    ])
-    writeFileSync(join(tempDir, 'image.png'), binaryContent)
-
-    const { ws, messages } = await createDiffTestServer()
-
-    send(ws, { type: 'get_diff' })
-    const result = await waitForMessage(messages, 'diff_result', 5000)
-
-    assert.equal(result.error, null)
-    const binaryFile = result.files.find(f => f.path === 'image.png')
-    assert.ok(binaryFile, 'Binary untracked file should be present')
-    assert.equal(binaryFile.status, 'untracked')
-    assert.equal(binaryFile.additions, 0, 'Binary file should have 0 additions')
-    assert.equal(binaryFile.hunks.length, 1)
-    assert.equal(binaryFile.hunks[0].lines.length, 1)
-    assert.equal(binaryFile.hunks[0].lines[0].type, 'context')
-    assert.ok(binaryFile.hunks[0].lines[0].content.includes('Binary file'), 'Should show binary placeholder')
-
-    ws.close()
-  })
-
-  it('returns error when no sessionCwd', async () => {
-    // Create a mock session without cwd set (cwd is undefined)
-    const mockSession = createMockSession()
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, true)
-
-    send(ws, { type: 'get_diff' })
-    const result = await waitForMessage(messages, 'diff_result', 5000)
-
-    assert.ok(result.error, 'Should return error when no CWD')
-    assert.match(result.error, /not available/i)
-
-    ws.close()
-  })
-})
-
-// ---------------------------------------------------------------------------
-// browse_files and read_file handler tests (#663)
-// ---------------------------------------------------------------------------
-describe('browse_files and read_file handlers', () => {
-  let server
-  let tempDir
-
-  beforeEach(() => {
-    // Resolve symlinks (macOS /tmp -> /private/tmp) so paths match CWD realpath checks
-    tempDir = realpathSync(mkdtempSync(join(tmpdir(), 'chroxy-fb-test-')))
-
-    // Build a test directory tree:
-    //   tempDir/
-    //     alpha.js
-    //     beta.py
-    //     .hidden
-    //     node_modules/
-    //       dep/
-    //     subdir/
-    //       nested.txt
-    //     zeta/
-    mkdirSync(join(tempDir, 'subdir'))
-    mkdirSync(join(tempDir, 'zeta'))
-    mkdirSync(join(tempDir, 'node_modules', 'dep'), { recursive: true })
-    writeFileSync(join(tempDir, 'alpha.js'), 'const a = 1')
-    writeFileSync(join(tempDir, 'beta.py'), 'print("hi")')
-    writeFileSync(join(tempDir, '.hidden'), 'secret')
-    writeFileSync(join(tempDir, 'subdir', 'nested.txt'), 'nested content')
-    writeFileSync(join(tempDir, 'node_modules', 'dep', 'index.js'), 'module.exports = {}')
-  })
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-    rmSync(tempDir, { recursive: true, force: true })
-  })
-
-  /** Spin up a WsServer with cwd set to tempDir and return a connected client. */
-  async function createTestServer(opts = {}) {
-    const mockSession = createMockSession()
-    if (opts.cwd !== undefined) {
-      mockSession.cwd = opts.cwd
-    } else {
-      mockSession.cwd = tempDir
-    }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port, true)
-    return { ws, messages }
-  }
-
-  // ------- browse_files -------
-
-  it('browse_files: lists files in session CWD', async () => {
-    const { ws, messages } = await createTestServer()
-
-    send(ws, { type: 'browse_files', path: '' })
-    const listing = await waitForMessage(messages, 'file_listing', 2000)
-
-    assert.equal(listing.error, null, 'Should not return an error')
-    assert.ok(listing.entries.length > 0, 'Should return entries')
-
-    // Check entries have expected shape
-    for (const entry of listing.entries) {
-      assert.equal(typeof entry.name, 'string')
-      assert.equal(typeof entry.isDirectory, 'boolean')
-      // size is null for directories, number for files
-      if (!entry.isDirectory) {
-        assert.equal(typeof entry.size, 'number')
-      }
-    }
-
-    // alpha.js should be present
-    assert.ok(listing.entries.some(e => e.name === 'alpha.js'), 'Should include alpha.js')
-    // subdir should be present
-    assert.ok(listing.entries.some(e => e.name === 'subdir' && e.isDirectory), 'Should include subdir/')
-
-    ws.close()
-  })
-
-  it('browse_files: sorts directories first, then alphabetical', async () => {
-    const { ws, messages } = await createTestServer()
-
-    send(ws, { type: 'browse_files', path: '' })
-    const listing = await waitForMessage(messages, 'file_listing', 2000)
-
-    assert.equal(listing.error, null)
-
-    const dirs = listing.entries.filter(e => e.isDirectory)
-    const files = listing.entries.filter(e => !e.isDirectory)
-
-    // All directories should come before all files
-    const lastDirIdx = listing.entries.lastIndexOf(dirs[dirs.length - 1])
-    const firstFileIdx = listing.entries.indexOf(files[0])
-    assert.ok(lastDirIdx < firstFileIdx, 'Directories should come before files')
-
-    // Directories should be alphabetical among themselves
-    for (let i = 1; i < dirs.length; i++) {
-      assert.ok(dirs[i - 1].name.localeCompare(dirs[i].name) <= 0,
-        `Dir ${dirs[i - 1].name} should come before ${dirs[i].name}`)
-    }
-
-    // Files should be alphabetical among themselves
-    for (let i = 1; i < files.length; i++) {
-      assert.ok(files[i - 1].name.localeCompare(files[i].name) <= 0,
-        `File ${files[i - 1].name} should come before ${files[i].name}`)
-    }
-
-    ws.close()
-  })
-
-  it('browse_files: filters dotfiles and node_modules', async () => {
-    const { ws, messages } = await createTestServer()
-
-    send(ws, { type: 'browse_files', path: '' })
-    const listing = await waitForMessage(messages, 'file_listing', 2000)
-
-    assert.equal(listing.error, null)
-
-    const names = listing.entries.map(e => e.name)
-    assert.ok(!names.includes('.hidden'), 'Should not include dotfiles')
-    assert.ok(!names.includes('node_modules'), 'Should not include node_modules')
-
-    ws.close()
-  })
-
-  it('browse_files: defaults to CWD when path is empty or null', async () => {
-    const { ws, messages } = await createTestServer()
-
-    // Test with empty string
-    send(ws, { type: 'browse_files', path: '' })
-    const listing1 = await waitForMessage(messages, 'file_listing', 2000)
-    assert.equal(listing1.error, null, 'Empty string should not error')
-    assert.ok(listing1.entries.length > 0, 'Should return entries for empty path')
-    const names1 = listing1.entries.map(e => e.name)
-
-    // Clear messages for next request
-    messages.length = 0
-
-    // Test with null
-    send(ws, { type: 'browse_files', path: null })
-    const listing2 = await waitForMessage(messages, 'file_listing', 2000)
-    assert.equal(listing2.error, null, 'Null path should not error')
-
-    // Both should return the same entries (CWD root)
-    const names2 = listing2.entries.map(e => e.name)
-    assert.deepEqual(names1, names2, 'Empty and null should return same entries')
-
-    ws.close()
-  })
-
-  it('browse_files: rejects path traversal outside CWD', async () => {
-    const { ws, messages } = await createTestServer()
-
-    send(ws, { type: 'browse_files', path: '../../etc' })
-    const listing = await waitForMessage(messages, 'file_listing', 2000)
-
-    assert.ok(listing.error, 'Should return an error for path traversal')
-    assert.match(listing.error, /access denied/i)
-    assert.deepEqual(listing.entries, [])
-
-    // Also test absolute paths outside CWD
-    messages.length = 0
-    send(ws, { type: 'browse_files', path: '/etc' })
-    const listing2 = await waitForMessage(messages, 'file_listing', 2000)
-
-    assert.ok(listing2.error, 'Should return an error for absolute path outside CWD')
-    assert.match(listing2.error, /access denied/i)
-    assert.deepEqual(listing2.entries, [])
-
-    ws.close()
-  })
-
-  it('browse_files: returns error when no session CWD', async () => {
-    const { ws, messages } = await createTestServer({ cwd: null })
-
-    send(ws, { type: 'browse_files', path: '' })
-    const listing = await waitForMessage(messages, 'file_listing', 2000)
-
-    assert.ok(listing.error, 'Should return an error when no CWD')
-    assert.match(listing.error, /not available/i)
-    assert.deepEqual(listing.entries, [])
-
-    ws.close()
-  })
-
-  it('browse_files: returns error for non-existent directory', async () => {
-    const { ws, messages } = await createTestServer()
-
-    send(ws, { type: 'browse_files', path: 'does-not-exist' })
-    const listing = await waitForMessage(messages, 'file_listing', 2000)
-
-    assert.ok(listing.error, 'Should return an error for non-existent directory')
-    assert.deepEqual(listing.entries, [])
-
-    ws.close()
-  })
-
-  // ------- read_file -------
-
-  it('read_file: reads a text file', async () => {
-    const { ws, messages } = await createTestServer()
-
-    send(ws, { type: 'read_file', path: 'alpha.js' })
-    const content = await waitForMessage(messages, 'file_content', 2000)
-
-    assert.equal(content.error, null, 'Should not return an error')
-    assert.equal(content.content, 'const a = 1')
-    assert.equal(typeof content.size, 'number')
-    assert.equal(content.truncated, false)
-
-    ws.close()
-  })
-
-  it('read_file: detects language from file extension', async () => {
-    const { ws, messages } = await createTestServer()
-
-    // .js -> js
-    send(ws, { type: 'read_file', path: 'alpha.js' })
-    const jsContent = await waitForMessage(messages, 'file_content', 2000)
-    assert.equal(jsContent.language, 'js', 'Should detect .js extension')
-
-    // .py -> py
-    messages.length = 0
-    send(ws, { type: 'read_file', path: 'beta.py' })
-    const pyContent = await waitForMessage(messages, 'file_content', 2000)
-    assert.equal(pyContent.language, 'py', 'Should detect .py extension')
-
-    // .txt -> txt
-    messages.length = 0
-    send(ws, { type: 'read_file', path: 'subdir/nested.txt' })
-    const txtContent = await waitForMessage(messages, 'file_content', 2000)
-    assert.equal(txtContent.language, 'txt', 'Should detect .txt extension')
-
-    ws.close()
-  })
-
-  it('read_file: rejects path traversal outside CWD', async () => {
-    const { ws, messages } = await createTestServer()
-
-    send(ws, { type: 'read_file', path: '../../etc/passwd' })
-    const content = await waitForMessage(messages, 'file_content', 2000)
-
-    assert.ok(content.error, 'Should return an error for path traversal')
-    assert.match(content.error, /access denied/i)
-    assert.equal(content.content, null)
-
-    // Also test absolute path outside CWD
-    messages.length = 0
-    send(ws, { type: 'read_file', path: '/etc/passwd' })
-    const content2 = await waitForMessage(messages, 'file_content', 2000)
-
-    assert.ok(content2.error, 'Should return an error for absolute path outside CWD')
-    assert.match(content2.error, /access denied/i)
-    assert.equal(content2.content, null)
-
-    ws.close()
-  })
-
-  it('read_file: rejects files over 512KB', async () => {
-    // Create a file slightly over 512KB
-    const largeContent = 'x'.repeat(512 * 1024 + 1)
-    writeFileSync(join(tempDir, 'large.bin'), largeContent)
-
-    const { ws, messages } = await createTestServer()
-
-    send(ws, { type: 'read_file', path: 'large.bin' })
-    const content = await waitForMessage(messages, 'file_content', 2000)
-
-    assert.ok(content.error, 'Should return an error for large file')
-    assert.match(content.error, /too large/i)
-    assert.equal(content.content, null)
-    assert.equal(typeof content.size, 'number')
-    assert.ok(content.size > 512 * 1024, 'Should report actual file size')
-
-    ws.close()
-  })
-
-  it('read_file: truncates content over 100KB', async () => {
-    // Create a file over 100KB but under 512KB
-    const bigContent = 'a'.repeat(150 * 1024)
-    writeFileSync(join(tempDir, 'big.txt'), bigContent)
-
-    const { ws, messages } = await createTestServer()
-
-    send(ws, { type: 'read_file', path: 'big.txt' })
-    const content = await waitForMessage(messages, 'file_content', 2000)
-
-    assert.equal(content.error, null, 'Should not return an error')
-    assert.equal(content.truncated, true, 'Should be marked as truncated')
-    assert.equal(content.content.length, 100 * 1024, 'Content should be truncated to 100KB')
-
-    ws.close()
-  })
-
-  it('read_file: detects binary files', async () => {
-    // Create a file with null bytes (binary)
-    const binaryContent = Buffer.alloc(100)
-    binaryContent[0] = 0x89  // PNG-like header
-    binaryContent[1] = 0x50
-    binaryContent[2] = 0x4e
-    binaryContent[3] = 0x47
-    binaryContent[10] = 0x00 // null byte
-    writeFileSync(join(tempDir, 'image.png'), binaryContent)
-
-    const { ws, messages } = await createTestServer()
-
-    send(ws, { type: 'read_file', path: 'image.png' })
-    const content = await waitForMessage(messages, 'file_content', 2000)
-
-    assert.ok(content.error, 'Should return an error for binary file')
-    assert.match(content.error, /binary/i)
-    assert.equal(content.content, null)
-
-    ws.close()
-  })
-
-  it('read_file: returns error for directories', async () => {
-    const { ws, messages } = await createTestServer()
-
-    send(ws, { type: 'read_file', path: 'subdir' })
-    const content = await waitForMessage(messages, 'file_content', 2000)
-
-    assert.ok(content.error, 'Should return an error for directory')
-    assert.match(content.error, /cannot read a directory/i)
-    assert.equal(content.content, null)
-
-    ws.close()
-  })
-
-  it('read_file: returns error for non-existent file', async () => {
-    const { ws, messages } = await createTestServer()
-
-    send(ws, { type: 'read_file', path: 'does-not-exist.txt' })
-    const content = await waitForMessage(messages, 'file_content', 2000)
-
-    assert.ok(content.error, 'Should return an error for non-existent file')
-    assert.match(content.error, /not found/i)
-    assert.equal(content.content, null)
-
-    ws.close()
-  })
-
-  it('read_file: returns error when no session CWD', async () => {
-    const { ws, messages } = await createTestServer({ cwd: null })
-
-    send(ws, { type: 'read_file', path: 'alpha.js' })
-    const content = await waitForMessage(messages, 'file_content', 2000)
-
-    assert.ok(content.error, 'Should return an error when no CWD')
-    assert.match(content.error, /not available/i)
-    assert.equal(content.content, null)
-
-    ws.close()
-  })
-})
 
 /** Create a mock SessionManager with configurable history */
 function createHistoryMockManager({ history = [], truncated = false, sessions = [] } = {}) {
@@ -5919,115 +2236,6 @@ describe('encryption integration (end-to-end)', () => {
     assert.ok(serverError.message.includes('key exchange timed out'), 'Error should mention key exchange timeout')
     assert.equal(serverError.recoverable, false, 'Error should be non-recoverable')
     assert.equal(closeCode, 1008, 'Server should close with code 1008 (policy violation)')
-  })
-})
-
-describe('models_updated broadcasting', () => {
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('broadcasts available_models when cliSession emits models_updated', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port)
-
-    // Wait for initial messages (auth_ok, server_mode, etc.)
-    await waitForMessage(messages, 'status')
-    const initialCount = messages.length
-
-    // Emit models_updated from the CLI session
-    const newModels = [
-      { id: 'sonnet-4-6', label: 'Sonnet 4.6', fullId: 'claude-sonnet-4-6' },
-      { id: 'opus-4-6', label: 'Opus 4.6', fullId: 'claude-opus-4-6' },
-    ]
-    mockSession.emit('models_updated', { models: newModels })
-
-    // Wait for broadcast (a NEW available_models after initialCount)
-    await waitFor(
-      () => messages.slice(initialCount).some(m => m.type === 'available_models'),
-      { label: 'available_models broadcast after emit' }
-    )
-
-    const modelMsgs = messages.slice(initialCount).filter(m => m.type === 'available_models')
-    assert.equal(modelMsgs.length, 1, 'Should receive exactly one available_models broadcast')
-    assert.deepStrictEqual(modelMsgs[0].models, newModels, 'Models should match emitted data')
-
-    ws.close()
-  })
-
-  it('broadcasts to all connected clients', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const client1 = await createClient(port)
-    const client2 = await createClient(port)
-
-    await waitForMessage(client1.messages, 'status')
-    await waitForMessage(client2.messages, 'status')
-    const c1Start = client1.messages.length
-    const c2Start = client2.messages.length
-
-    const newModels = [{ id: 'test-model', label: 'Test', fullId: 'claude-test-model' }]
-    mockSession.emit('models_updated', { models: newModels })
-
-    await waitFor(
-      () => client1.messages.slice(c1Start).some(m => m.type === 'available_models'),
-      { label: 'c1 available_models after emit' }
-    )
-    await waitFor(
-      () => client2.messages.slice(c2Start).some(m => m.type === 'available_models'),
-      { label: 'c2 available_models after emit' }
-    )
-
-    const c1Models = client1.messages.slice(c1Start).filter(m => m.type === 'available_models')
-    const c2Models = client2.messages.slice(c2Start).filter(m => m.type === 'available_models')
-    assert.equal(c1Models.length, 1, 'Client 1 should receive models update')
-    assert.equal(c2Models.length, 1, 'Client 2 should receive models update')
-
-    client1.ws.close()
-    client2.ws.close()
-  })
-
-  it('ignores models_updated with missing models data', async () => {
-    const mockSession = createMockSession()
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-    const { ws, messages } = await createClient(port)
-
-    await waitForMessage(messages, 'status')
-    const initialCount = messages.length
-
-    // Emit with no models property
-    mockSession.emit('models_updated', {})
-    // Negative test: wait a short time to confirm no message arrives
-    await new Promise(r => setTimeout(r, 100))
-
-    const modelMsgs = messages.slice(initialCount).filter(m => m.type === 'available_models')
-    assert.equal(modelMsgs.length, 0, 'Should not broadcast when models data is missing')
-
-    ws.close()
   })
 })
 
@@ -7429,124 +3637,6 @@ describe('provider capability gates', () => {
   })
 })
 
-describe('client_focus_changed broadcast', () => {
-  let server
-
-  afterEach(() => {
-    if (server) {
-      server.close()
-      server = null
-    }
-  })
-
-  it('broadcasts client_focus_changed when a client switches session', async () => {
-    const { manager: mockManager } = createMockSessionManager([
-      { id: 'sess-a', name: 'Session A', cwd: '/tmp/a' },
-      { id: 'sess-b', name: 'Session B', cwd: '/tmp/b' },
-    ])
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    // Connect two clients
-    const client1 = await createClient(port, true)
-    const client2 = await createClient(port, true)
-    await waitForMessage(client1.messages, 'auth_ok', 2000)
-    await waitForMessage(client2.messages, 'auth_ok', 2000)
-
-    // Client 1 switches to session B
-    send(client1.ws, { type: 'switch_session', sessionId: 'sess-b' })
-
-    // Client 2 should receive client_focus_changed
-    const focusMsg = await waitForMessage(client2.messages, 'client_focus_changed', 2000)
-    assert.ok(focusMsg, 'Client 2 should receive client_focus_changed')
-    assert.equal(focusMsg.sessionId, 'sess-b', 'Should indicate the new session')
-    assert.equal(typeof focusMsg.clientId, 'string', 'Should include clientId')
-    assert.equal(typeof focusMsg.timestamp, 'number', 'Should include timestamp')
-
-    client1.ws.close()
-    client2.ws.close()
-  })
-
-  it('does not send client_focus_changed to the switching client itself', async () => {
-    const { manager: mockManager } = createMockSessionManager([
-      { id: 'sess-a', name: 'Session A', cwd: '/tmp/a' },
-      { id: 'sess-b', name: 'Session B', cwd: '/tmp/b' },
-    ])
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const client1 = await createClient(port, true)
-    await waitForMessage(client1.messages, 'auth_ok', 2000)
-
-    send(client1.ws, { type: 'switch_session', sessionId: 'sess-b' })
-    await waitForMessage(client1.messages, 'session_switched', 2000)
-
-    // Verify no client_focus_changed is delivered to the switching client within a reasonable timeout
-    try {
-      await waitForMessage(client1.messages, 'client_focus_changed', 500)
-      assert.fail('Switching client should NOT receive client_focus_changed')
-    } catch {
-      // Expected: waitForMessage times out because no client_focus_changed is sent to the switcher
-    }
-
-    client1.ws.close()
-  })
-
-  it('broadcasts client_focus_changed when a client creates a session', async () => {
-    const { manager: mockManager, sessionsMap } = createMockSessionManager([
-      { id: 'sess-a', name: 'Session A', cwd: '/tmp/a' },
-    ])
-
-    // Add createSession to mock — inserts into sessionsMap so getSession/listSessions/firstSessionId all reflect it
-    let nextId = 1
-    mockManager.createSession = ({ name, cwd }) => {
-      const id = `new-sess-${nextId++}`
-      const mockSession = createMockSession()
-      mockSession.cwd = cwd || '/tmp'
-      sessionsMap.set(id, { session: mockSession, name: name || 'New Session', cwd: cwd || '/tmp', type: 'cli', isBusy: false })
-      return id
-    }
-
-    server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      sessionManager: mockManager,
-      authRequired: false,
-    })
-    const port = await startServerAndGetPort(server)
-
-    const client1 = await createClient(port, true)
-    const client2 = await createClient(port, true)
-    await waitForMessage(client1.messages, 'auth_ok', 2000)
-    await waitForMessage(client2.messages, 'auth_ok', 2000)
-
-    // Client 1 creates a new session
-    send(client1.ws, { type: 'create_session', name: 'My Session' })
-
-    // Client 2 should receive client_focus_changed
-    const focusMsg = await waitForMessage(client2.messages, 'client_focus_changed', 2000)
-    assert.ok(focusMsg, 'Client 2 should receive client_focus_changed')
-    assert.ok(focusMsg.sessionId.startsWith('new-sess-'), 'Should reference the new session')
-    assert.equal(typeof focusMsg.clientId, 'string', 'Should include clientId')
-    assert.equal(typeof focusMsg.timestamp, 'number', 'Should include timestamp')
-
-    client1.ws.close()
-    client2.ws.close()
-  })
-})
-
 describe('_sendSessionInfo sessionId tagging (#1417)', () => {
   let server
 
@@ -8061,54 +4151,3 @@ describe('CORS origin restrictions (#1533)', () => {
   })
 })
 
-describe('broadcast backpressure', () => {
-  it('skips clients whose bufferedAmount exceeds the threshold', async () => {
-    const mockSession = createMockSession()
-    const server = new WsServer({
-      port: 0,
-      apiToken: 'test-token',
-      cliSession: mockSession,
-      authRequired: false,
-      backpressureThreshold: 100,
-    })
-    // Clear log listener to prevent log_entry broadcasts from interfering
-    // with backpressure threshold measurements (#1820)
-    const { setLogListener } = await import('../src/logger.js')
-    setLogListener(null)
-    const port = await startServerAndGetPort(server)
-
-    const { ws, messages } = await createClient(port, true)
-
-    // Verify normal broadcast works
-    server.broadcast({ type: 'discovered_sessions', tmux: [] })
-    await waitFor(() => messages.find(m => m.type === 'discovered_sessions'), { label: 'normal broadcast' })
-    assert.ok(messages.find(m => m.type === 'discovered_sessions'), 'Should receive broadcast when under threshold')
-
-    // Stub bufferedAmount to simulate backpressure
-    const clientWs = [...server.clients.keys()][0]
-    Object.defineProperty(clientWs, 'bufferedAmount', { get: () => 200, configurable: true })
-
-    // Spy on _send to verify backpressure skips the client
-    const sendCalls = []
-    const originalSend = server._send.bind(server)
-    server._send = (ws, msg) => { sendCalls.push({ ws, msg }); return originalSend(ws, msg) }
-
-    server.broadcast({ type: 'discovered_sessions', tmux: [{ sessionName: 'bp-test' }] })
-
-    const clientSends = sendCalls.filter(call => call.ws === clientWs)
-    assert.equal(clientSends.length, 0, 'Should NOT send broadcast to client when over backpressure threshold')
-    server._send = originalSend
-
-    // Restore bufferedAmount and verify broadcast resumes
-    Object.defineProperty(clientWs, 'bufferedAmount', { get: () => 0, configurable: true })
-    server.broadcast({ type: 'discovered_sessions', tmux: [{ sessionName: 'resumed' }] })
-    const resumedMsg = await waitFor(
-      () => messages.find(m => m.type === 'discovered_sessions' && m.tmux?.[0]?.sessionName === 'resumed'),
-      { label: 'resumed broadcast' }
-    )
-    assert.ok(resumedMsg, 'Should receive broadcast after backpressure clears')
-
-    ws.close()
-    server.close()
-  })
-})
